### 2025-05-17 #8224 @editorialbot
[PRE REVIEW]: wristpy: Fast, User-Friendly Python Processing of Wrist-worn Accelerometer Data
**Submitting author:** @Asanto32 (<a href="http://orcid.org/0000-0002-8951-4619">Adam Santorelli</a>)
**Repository:** https://github.com/childmindresearch/wristpy
**Branch with paper.md** (empty if default branch): paper/JOSS-submission
**Version:** 0.1.1
