### 2025-05-17 #8224 @Kevin-Mattheus-Moerman
@sappelhoff would this one be easy to handle for you as a +1 since you are handling another "Accelerometer" related package and there may be a similar reviewer pool? No worries if you are too busy too.
