### 2025-05-17 #8224 @sappelhoff
@Kevin-Mattheus-Moerman I cannot take this over right now, but perhaps in a month, if nobody has been found until then.

One thing I would recommend discussing is a topic similar to the one raised in the PAAT submission: Why are there so many accelerometer analysis packages, when a large part of the functionality seems to be overlapping (this is a statement only based on a first glance, I am very happy to stand corrected). See this thread and the responses below: https://github.com/openjournals/joss-reviews/issues/7817#issuecomment-2781465880
