### 2025-05-17 4f635c7
Update paper.bib
Update to main
Fix DOIs
