+++ paper.bib
@@ -1,7 +1,4 @@
 
-@unpublished{noauthor_notitle_nodate,
-}
-
 @misc{van_hees_ggir_2025,
 	title = {{GGIR}},
 	copyright = {Apache License 2.0},
@@ -106,8 +103,6 @@
 	file = {Full Text:/Users/adam.santorelli/Zotero/storage/2ZKV5GCF/Christakis et al. - 2019 - SleepPy A python package for sleep analysis from accelerometer data.pdf:application/pdf},
 }
 
-@misc{noauthor_notitle_nodate-1,
-}
 
 @article{john_open-source_2019,
 	title = {An {Open}-{Source} {Monitor}-{Independent} {Movement} {Summary} for {Accelerometer} {Data} {Processing}},
@@ -211,21 +206,6 @@ Full Changelog: https://github.com/childmindresearch/actfast/compare/v1.1.1...v1
 	file = {Full Text:/Users/adam.santorelli/Zotero/storage/ELG2LPGF/Hammad et al. - 2021 - pyActigraphy Open-source python package for actigraphy data visualization and analysis.pdf:application/pdf},
 }
 
-@article{hildebrand_age_2014-1,
-	title = {Age {Group} {Comparability} of {Raw} {Accelerometer} {Output} from {Wrist}- and {Hip}-{Worn} {Monitors}},
-	volume = {46},
-	issn = {0195-9131},
-	url = {https://journals.lww.com/00005768-201409000-00017},
-	doi = {10.1249/MSS.0000000000000289},
-	language = {en},
-	number = {9},
-	urldate = {2025-01-23},
-	journal = {Medicine \& Science in Sports \& Exercise},
-	author = {Hildebrand, Maria and Van Hees, Vincent T. and Hansen, Bjorge Hermann and Ekelund, Ulf},
-	month = sep,
-	year = {2014},
-	pages = {1816--1824},
-}
 
 @article{vert_detecting_2022,
 	title = {Detecting accelerometer non-wear periods using change in acceleration combined with rate-of-change in temperature},
