+++ tests/unit/test_metrics.py
@@ -70,13 +70,13 @@ def test_euclidean_norm_minus_one(
     expected_length = math.ceil(len(test_time) / 5)
     enmo_results = metrics.euclidean_norm_minus_one(test_acceleration)
 
-    assert np.all(
-        np.isclose(enmo_results.measurements, expected_enmo)
-    ), f"Expected {expected_enmo}"
+    assert np.all(np.isclose(enmo_results.measurements, expected_enmo)), (
+        f"Expected {expected_enmo}"
+    )
 
-    assert (
-        len(enmo_results.time) == expected_length
-    ), "Input time attribute does not match output time attribute."
+    assert len(enmo_results.time) == expected_length, (
+        "Input time attribute does not match output time attribute."
+    )
 
 
 @pytest.mark.parametrize(
@@ -106,9 +106,9 @@ def test_angle_relative_to_horizontal(
         np.isclose(angle_z_results.measurements, expected_anglez, equal_nan=True)
     ), f"Expected {expected_anglez}, got: {angle_z_results.measurements}"
 
-    assert (
-        len(angle_z_results.time) == expected_length
-    ), "Input time attribute does not match output time attribute."
+    assert len(angle_z_results.time) == expected_length, (
+        "Input time attribute does not match output time attribute."
+    )
 
 
 @pytest.mark.parametrize(
@@ -130,9 +130,9 @@ def test_cleanup_isolated_ones_nonwear_value(
     """Test the cleanup isolated ones nonwear value function."""
     test_result = metrics._cleanup_isolated_ones_nonwear_value(nonwear_value_array)
 
-    assert np.all(
-        test_result == expected_result
-    ), f"Expected {expected_result}, got: {test_result}"
+    assert np.all(test_result == expected_result), (
+        f"Expected {expected_result}, got: {test_result}"
+    )
 
 
 def test_group_acceleration_data_by_time() -> None:
@@ -173,9 +173,9 @@ def test_compute_nonwear_value_per_axis(
         acceleration_grouped["X"], std_criteria=modifier
     )
 
-    assert (
-        test_resultx == expected_result
-    ), f"Expected {expected_result}, got: {test_resultx}"
+    assert test_resultx == expected_result, (
+        f"Expected {expected_result}, got: {test_resultx}"
+    )
 
 
 def test_compute_nonwear_value_array(create_acceleration: pl.DataFrame) -> None:
@@ -194,12 +194,12 @@ def test_compute_nonwear_value_array(create_acceleration: pl.DataFrame) -> None:
         std_criteria=1,
     )
 
-    assert np.all(
-        test_result == expected_result
-    ), f"Expected {expected_result}, got: {test_result}"
-    assert (
-        len(test_result) == expected_time_length
-    ), f"Expected time to be {expected_time_length}, got: {len(test_result)}"
+    assert np.all(test_result == expected_result), (
+        f"Expected {expected_result}, got: {test_result}"
+    )
+    assert len(test_result) == expected_time_length, (
+        f"Expected time to be {expected_time_length}, got: {len(test_result)}"
+    )
 
 
 @pytest.mark.parametrize(
@@ -229,12 +229,12 @@ def test_detect_nonwear(
         std_criteria=modifier,
     )
 
-    assert np.all(
-        test_result.measurements == modifier
-    ), f"Expected non-wear flag value to be {expected_result}, got: {test_result}"
-    assert (
-        len(test_result.time) == expected_time_length
-    ), f"Expected time to be {expected_time_length}, got: {len(test_result.time)}"
+    assert np.all(test_result.measurements == modifier), (
+        f"Expected non-wear flag value to be {expected_result}, got: {test_result}"
[…]
