### 2025-05-20 #8224 @Asanto32
Hi @sappelhoff
Thank you for the questions. That is a great comment and we agree that there appears to be a growing number of open source tools for processing accelerometer data. In fact, the development of `wristpy` stems from our internal, and collaborators, use of the GGIR package (the most widely used package for accelerometer processing for both sleep and physical activity analysis). We conducted a needs-based analysis to find what specifically users (both internal and collaborators) would want in a tool, and it was with these needs in mind that we developed `wristpy`.

 We agree that there is significant overlap between many of these tools (GGIR, `skdh`, `PAAT`, and `wristpy`). We strived to build off the best parts of some of these packages; however, and we aimed to highlight this in the paper, what makes `wristpy` stand apart is that we can process both raw .gt3x (Actigraph) and .bin (GENEActiv) files and include any number of sleep windows; additionally, we use our own in-house built rust-based reader that allows us to process these large files in an order of magnitude faster than other packages. Furthermore, we have strived to develop a truly modular package that is extensively documented, has an easily readable and understandable codebase, allows for various interfaces through a CLI, Python entrypoint, and a Docker image, and has exhaustive testing to allow for extensible development from both our developers and the community.

Throughout the development process we have kept the end-user in mind. In fact, it was through discussions with KOLs in the field that we identified the most important features to add to `wristpy` for the current version release. We hope that this allows the package to be useful both internally (we have already started to use it to process our large HBN actigraphy dataset) and for current and future collaborators. Our aim is for `wristpy` to primarily be a processing package, as we look to further develop special use-case extendible tools for specific post-processing needs (such as the sleep analysis toolbox we are currently working on).

Hope this helps! Happy to answer any other questions. We believe that this toolbox can be very helpful and beneficial to the community!
