+++ tests/unit/test_readers.py
@@ -61,11 +61,11 @@ def test_extract_dynamic_range_bin(sample_data_bin: pathlib.Path) -> None:
     ), f"Expected dynamic range of: {expected_dynamic_range}, result was: {result}"
 
 
-def test_extract_dynamic_range_gt3x(sample_data_bin: pathlib.Path) -> None:
+def test_extract_dynamic_range_gt3x(sample_data_gt3x: pathlib.Path) -> None:
     """Test extracting dynamic range metadata from .gt3x files."""
     expected_dynamic_range = (-8, 8)
-    data = actfast.read(sample_data_bin)
-    file_type = cast(Literal[".gt3x", ".bin"], sample_data_bin.suffix)
+    data = actfast.read(sample_data_gt3x)
+    file_type = cast(Literal[".gt3x", ".bin"], sample_data_gt3x.suffix)
 
     result = readers._extract_dynamic_range(
         metadata=data["metadata"], file_type=file_type
