### 2025-05-22 b1340a0
Initial refactor of output
Moved OrchestratorResults model to io/writers
We now output a .json file with the config parameters
