+++ tests/smoke/test_orchestrator_smoke.py
@@ -5,6 +5,7 @@ import pathlib
 import pytest
 
 from wristpy.core import models, orchestrator
+from wristpy.io.writers import writers
 
 
 @pytest.mark.parametrize(
@@ -21,7 +22,7 @@ def test_orchestrator_happy_path(
     results = orchestrator._run_file(input=sample_data_gt3x, output=output_path)
 
     assert output_path.exists()
-    assert isinstance(results, models.OrchestratorResults)
+    assert isinstance(results, writers.OrchestratorResults)
     assert isinstance(results.physical_activity_metric, models.Measurement)
     assert isinstance(results.anglez, models.Measurement)
     assert isinstance(results.nonwear_epoch, models.Measurement)
@@ -39,7 +40,7 @@ def test_orchestrator_idle_sleep_mode_run(
     )
 
     assert (tmp_path / "good_file.csv").exists()
-    assert isinstance(results, models.OrchestratorResults)
+    assert isinstance(results, writers.OrchestratorResults)
     assert isinstance(results.physical_activity_metric, models.Measurement)
     assert isinstance(results.anglez, models.Measurement)
     assert isinstance(results.nonwear_epoch, models.Measurement)
