### 2025-05-23 762a8c8
Update test_orchestrator_smoke.py
Simple test
