### 2025-05-23 959b2fb
Update test_orchestrator_smoke.py
Fix typo of default calibrator...
