### 2025-05-23 9fbd35b
Create test_writers.py
Test empty params
