### 2025-05-23 833b9b3
This is covered by `validate_output`
