+++ tests/unit/test_writers.py
@@ -2,7 +2,6 @@
 
 import datetime
 import logging
-
 import pathlib
 
 import numpy as np
