+++ README.md
@@ -143,7 +143,7 @@ sleep_windows = subject1.sleep_windows_epoch
     docker run -it --rm \
      -v "/local/path/to/data/file.bin:/data/file.bin" \
      -v "/local/path/to/output:/output" \
-     cmidair/wristpy
+     cmidair/wristpy:main
    ```
 
 ### Customizing the Pipeline:
@@ -154,7 +154,7 @@ The Docker image supports multiple input variables to customize processing. You 
 docker run -it --rm \
   -v "/local/path/to/data/file.bin:/data/file.bin" \
   -v "/local/path/to/output:/output" \
-  cmidair/wristpy /data --output /output --epoch-length 5 --nonwear-algorithm ggir --nonwear-algorithm detach --thresholds 0.1 0.2 0.4
+  cmidair/wristpy:main /data --output /output --epoch-length 5 --nonwear-algorithm ggir --nonwear-algorithm detach --thresholds 0.1 0.2 0.4
 ```
 
 
