### 2025-05-30 8989e6d
Update metrics.py
Ruff line length
