### 2025-06-02 6e09e35
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
