+++ .github/workflows/test.yaml
@@ -13,7 +13,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        python_version: ['3.10', '3.11', '3.12','3.13']
+        python_version: ['3.10', '3.11', '3.12', '3.13']
         resolution: [lowest-direct, highest]
     steps:
     - uses: actions/checkout@v4
