### 2025-06-03 17c947a
Update README.md
Adding block note for macOS
Adding uv installation instructions
