### 2025-06-04 0e43d01
Update writers.py
One day I will run the pre-commit hooks locally...
