### 2025-06-04 b5167c4
Update README.md
Add line about wristpy --help
