### 2025-06-04 d1136e5
Changes from meeting
Clearly define 3.13 not support in pyproject.toml
Change logger level + handling of verbosity
