### 2025-06-04 dccb0f4
Adding logger info, suppressing polars warning
