### 2025-06-05 147c925
Update orchestrator.py
Fixing logging level information for when files/directory completed successfully.
