### 2025-06-05 b6950df
Update analytics.py
Removed line that lists all SIB and SPT windows (typically hundreds)
