### 2025-06-06 daba59c
Update test_orchestrator.py
Fixed mims test for new writers output
