### 2025-06-09 cece6b1
using callback to run version and exit.
