### 2025-06-09 4379f89
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
