### 2025-06-09 e4db668
reverting to using callback in option.
