### 2025-06-10 91587a6
Feature/issue 200/update output (#201)
* Initial refactor of output

Moved OrchestratorResults model to io/writers
We now output a .json file with the config parameters

* Update test_orchestrator_smoke.py

Simple test

* Update test_orchestrator_smoke.py

Fix typo of default calibrator...

* Update test_orchestrator_smoke.py

trying to push actions

* Renaming attributes for csv export

* Create test_writers.py

Test empty params

* This is covered by `validate_output`

* Update README.md

Fixing missing readme update from previous PR

* Fixing typos

* Update README.md

Docker tag missing

*
[…]
