### 2025-06-10 982f06e
Bump the production-dependencies group across 1 directory with 2 updates
Bumps the production-dependencies group with 2 updates in the / directory: [pyarrow](https://github.com/apache/arrow) and [typer](https://github.com/fastapi/typer).


Updates `pyarrow` from 19.0.1 to 20.0.0
- [Release notes](https://github.com/apache/arrow/releases)
- [Commits](https://github.com/apache/arrow/compare/apache-arrow-19.0.1...apache-arrow-20.0.0)

Updates `typer` from 0.15.4 to 0.16.0
- [Release notes](https://github.com/fastapi/typer/releases)
- [Changelog](https://github.com/fastapi/typer/blob/master/docs/r
[…]
