+++ pyproject.toml
@@ -17,14 +17,14 @@ packages = [{include = "wristpy", from = "src"}]
 python = ">=3.10, <3.13"
 polars = "^1.21.0"
 scikit-learn = "^1.4.1.post1"
-pyarrow = ">=18,<20"
+pyarrow = ">=18,<21"
 actfast = "^1.1.0"
 numpy = ">=1.26.0, <3.0.0"
 pydantic = "^2.7.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.13.0"
 scikit-digital-health = "^0.17.3"
-typer = "^0.15.2"
+typer = ">=0.15.2,<0.17.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.0"
