### 2025-06-11 childmindresearch/wristpy#218 @Asanto32
Task: Replace tutorial with sphinx
### Description

As the tutorial and documentation has started to grow we will migrate existing documentation to sphinx

### Tasks

- [ ] Task 1

### Freeform Notes

_No response_
