### 2025-06-12 5a4e51c
203 bug checking version (#215)
* version argument change to is_eager.

* using callback to run version and exit.

* reverting to using callback in option.

* testing

* removing duplicate code.

* typo.
