+++ docs/wristpy_tutorial.md
@@ -1,22 +1,24 @@
-Wristpy Tutorial
-================
+# Wristpy Tutorial
 
-Introduction
-------------
+
+## Introduction
+
 
 Wristpy is a Python library designed for processing and analyzing wrist-worn accelerometer data.
-This tutorial will guide you through the basic steps of using Wristpy to analyze your accelerometer data. Specifically,
+This tutorial will guide you through the basic steps of using `wristpy` to analyze your accelerometer data. Specifically,
 we will cover the following topics through a few examples:
    - running the default processor, analyzing the output data, and visualizing the results.
    - loading data and plotting the raw signals.
    - how to calibrate the data, computing ENMO and angle-z from the calibrated data and then plotting those metrics.
    - how to obtain non-wear windows and visualize them.
    - how to obtain sleep windows and visualize them.
-      - how we can filter sleep windows that overlap with non-wear periods.
+      - how we can filter sleep windows that overlap with non-wear periods.   
+  
+    
 
 
-Example 1: Running the default processor
-----------------------------------------
+## Example 1: Running the default processor
+
 
 The `orchestrator` module of wristpy contains the default processor that will run the entire wristpy processing pipeline. This can be called as simply as:
 
@@ -28,7 +30,8 @@ results = orchestrator.run(
    output = 'path/to/save/file_name.csv'
 )
 ```
-This runs the processing pipeline with all the default arguments, creates an output `.csv` file, and will create a `results` object that contains the various output metrics (namely, enmo, angle-z, physical activity values, non-wear detection, sleep detection).
+This runs the processing pipeline with all the default arguments, creates an output `.csv` file, a `.json` file with the pipeline configuration parameters, and will create a `results` object that contains the various output metrics (namely, the specified physical activity metric, angle-z, physical activity classification values, non-wear detection, sleep detection).
+
 
 The orchestrator can also process entire directories. The call to the orchestrator remains largely the same but now output is expected to be a directory and the desired filetype for the saved files **must** be specified:
 
@@ -49,7 +52,7 @@ We can visualize some of the outputs within the `results` object, directly, with
 Plot the ENMO across the entire data set:
 ```python
 from matplotlib import pyplot as plt
-plt.plot(results.enmo.time, results.enmo.measurements)
+plt.plot(results.physical_activity_metric.time, results.physical_activity_metric.measurements)
 ```
 
 ![Example of the ENMO result](enmo_example1.png)
@@ -59,7 +62,7 @@ Plot the sleep windows with normalized angle-z data:
 from matplotlib import pyplot as plt
 
 plt.plot(results.anglez.time, results.anglez.measurements/90)
-plt.plot(results.sleep_windows_epoch.time, results.sleep_windows_epoch.measurements)
+plt.plot(results.sleep_status.time, results.sleep_status.measurements)
 plt.legend(['Angle Z', 'Sleep Windows'])
 plt.show()
 ```
@@ -80,10 +83,10 @@ plt.plot(output_results['time'], phys_activity)
 It is also possible to do some analysis on these output variables, for example, if we want to find the percent of time spent inactive, or in light, moderate, or vigorous physical activity:
 
 ```python
-inactivity_count = sum(output_results['physical_activity_levels'] == 0)
-light_activity_count = sum(output_results['physical_activity_levels'] == 1)
-moderate_activity_count = sum(output_results['physical_activity_levels'] == 2)
-vigorous_activity_count = sum(output_results['physical_activity_levels'] == 3)
+inactivity_count = sum(phys_activity == 0)
+light_activity_count = sum(phys_activity== 1)
[…]
