+++ docs/wristpy_tutorial.md
@@ -30,7 +30,7 @@ results = orchestrator.run(
    output = 'path/to/save/file_name.csv'
 )
 ```
-This runs the processing pipeline with all the default arguments, creates an output `.csv` file, a `.json` file with the pipeline configuration parameters, and will create a `results` object that contains the various output metrics (namely, the specified physical activity metric, angle-z, physical activity classification values, non-wear detection, sleep detection).
+This runs the processing pipeline with all the default arguments, creates an output `.csv` file, a `.json` file with the pipeline configuration parameters, and will create a `results` object that contains the various output metrics (namely; the specified physical activity metric, angle-z, physical activity classification values, non-wear status, and sleep status).
 
 
 The orchestrator can also process entire directories. The call to the orchestrator remains largely the same but now output is expected to be a directory and the desired filetype for the saved files **must** be specified:
@@ -49,7 +49,7 @@ results = orchestrator.run(
 
 We can visualize some of the outputs within the `results` object, directly, with the following scripts:
 
-Plot the ENMO across the entire data set:
+Plot the default physical activity metrics (ENMO) across the entire data set:
 ```python
 from matplotlib import pyplot as plt
 plt.plot(results.physical_activity_metric.time, results.physical_activity_metric.measurements)
@@ -104,9 +104,13 @@ Inactivity percent: 87.21338876269661
 
 > ### Configuring a custom pipeline
 >
-> We can modify the defualt values of the following parameters: calibration, epoch_length, physical activity metric, non-wear algorithm
->
->
+> A custom processing pipeline can be easily created by modifying the input arguments to the `orchestrator.run` call.
+> 
+>> For example:
+>> ```python
+>> results = orchestrator.run(input = '/path/to/input/dir', output = '/path/to/output/dir', output_filetype = ".parquet", calibrator="gradient", activity_metric="ag_count", nonwear_algorithm=["detach"], epoch_length=10, thresholds=[0.05, 0.1, 0.3])
+>> ```
+> Complete documentation on these parameters can be found [here](https://childmindresearch.github.io/wristpy/wristpy/core/orchestrator.html#run).
 
 
 ## Example 2: Loading data and plotting the raw signals
@@ -138,9 +142,9 @@ Plot the light data:
 
 
 
-## Example 3:  Plot the epoch1 level measurements
+## Example 3:  Plot the epoch-level measurements
 
-In this example we will expand on the skills learned in `Example 2`: we will load the sensor data, calibrate, and then calculate the ENMO and angle-z data in 5s windows (epoch 1 data).
+In this example we will expand on the skills learned in `Example 2`: we will load the sensor data, calibrate, and then calculate the ENMO and angle-z data in 5s windows (epoch-level data).
 
 ```python
 from wristpy.io.readers import readers
@@ -157,7 +161,7 @@ calibrated_data = calibrator_object.run_calibration(watch_data.acceleration)
 enmo = metrics.euclidean_norm_minus_one(calibrated_data)
 anglez = metrics.angle_relative_to_horizontal(calibrated_data)
 
-#Obtain the epoch1 level data
+#Obtain the epoch-level data, default is 5s windows
 enmo_epoch1 = computations.moving_mean(enmo)
 anglez_epoch1 = computations.moving_mean(anglez)
 ```
@@ -193,12 +197,12 @@ watch_data = readers.read_watch_data('/path/to/geneactive/file.bin')
 calibrator_object = calibration.ConstrainedMinimizationCalibration()
[…]
