### 2025-06-16 1df9af6
Update dependency minimum
