+++ pyproject.toml
@@ -20,7 +20,7 @@ scikit-learn = "^1.4.1.post1"
 pyarrow = ">=18,<20"
 actfast = "^1.1.0"
 numpy = ">=1.26.0, <3.0.0"
-pydantic = "^2.7.1"
+pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.13.0"
 scikit-digital-health = "^0.17.9"
