### 2025-06-16 7b40425
Update analytics.py
Cast signed np.int64 to int
