### 2025-06-16 fd44f80
Update nonwear image
