### 2025-06-17 #8224 @Asanto32
>
> [@Asanto32](https://github.com/Asanto32) do you have suggestions for potential reviewers? If so, please name their GitHub user names (but don't tag them). Please keep in mind the JOSS COI policy, too: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy

Hi,
Great, this is good timing in some ways as we are about to release a new version of `wristpy`.
The only reviewer I can think of that would be very relevant (and I know he is currently a reviewer for the PAAT paper) is: LukasAdamowicz

Thanks,
Adam
