### 2025-06-18 #8224 @sappelhoff
@LukasAdamowicz would you be willing to also review this one?
