### 2025-06-18 ac213cf
Bump the development-dependencies group across 1 directory with 3 updates
Bumps the development-dependencies group with 3 updates in the / directory: [pytest](https://github.com/pytest-dev/pytest), [mypy](https://github.com/python/mypy) and [ruff](https://github.com/astral-sh/ruff).


Updates `pytest` from 8.4.0 to 8.4.1
- [Release notes](https://github.com/pytest-dev/pytest/releases)
- [Changelog](https://github.com/pytest-dev/pytest/blob/main/CHANGELOG.rst)
- [Commits](https://github.com/pytest-dev/pytest/compare/8.4.0...8.4.1)

Updates `mypy` from 1.16.0 to 1.16.1
- [Changelog](https://gith
[…]
