### 2025-06-18 a27062c
Update metrics.py
