### 2025-06-18 9ff7209
Update wristpy_tutorial.md
