+++ docs/wristpy_tutorial.md
@@ -223,7 +223,7 @@ plt.legend(['ENMO Epoch1', 'Non-wear'])
 
 
 We can visualize the sleep periods in comparison to other metrics; in this example, we compare the sleep windows to the angle-z data and the non-wear periods. In the default pipeline any sleep periods that overlap with non-wear periods are filtered out.
-This plot shows the sleep periods visualized by a blue trace, non-wear periods are visualized with a green trace, and the angle-z data with the semi-transparent red trace. These are all accessible directly from the results object calculated from the default pipeline:
+This plot shows the sleep periods visualized by a blue trace, non-wear periods are visualized with a green trace, and the angle-z data with the semi-transparent red trace. These are all accessible directly from the results object created with the custom pipeline:
 
 ```python
 import matplotlib.pyplot as plt
