### 2025-06-18 ea55c94
Update wristpy_tutorial.md
Fix issue with potential conversion to categorical/numeric
