### 2025-06-18 e027220
Feature/issue 159/new version release (#219)
* Prepare metadata

* Update analytics.py

Removed line that lists all SIB and SPT windows (typically hundreds)

* Fixing typos

* Update README.md

* More typos!

* Updating tutorial (partially)

* Updating tutorial

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update README.md

* Update nonwear image

* Some more corrections

* Update metrics.py

* Update wristpy_tutorial.md

* Update wristpy_tutorial.md

Fix issue with potential conversion to categorical/numeric

---------

Co-authored
[…]
