### 2025-06-18 fb3f91f
Revert "Feature/issue 159/new version release (#219)"
This reverts commit e027220f93c483b886f39e47164749272ec186fe.
