+++ src/wristpy/io/writers/writers.py
@@ -1,4 +1,4 @@
-"""Module containing the output classes for writing data to files."""
+"""Module containg the output classes for writing data to files."""
 
 import datetime
 import json
