+++ docs/wristpy_tutorial.md
@@ -1,22 +1,24 @@
-Wristpy Tutorial
-================
+# Wristpy Tutorial
 
-Introduction
-------------
+
+## Introduction
+
 
 Wristpy is a Python library designed for processing and analyzing wrist-worn accelerometer data.
-This tutorial will guide you through the basic steps of using Wristpy to analyze your accelerometer data. Specifically,
+This tutorial will guide you through the basic steps of using `wristpy` to analyze your accelerometer data. Specifically,
 we will cover the following topics through a few examples:
    - running the default processor, analyzing the output data, and visualizing the results.
    - loading data and plotting the raw signals.
    - how to calibrate the data, computing ENMO and angle-z from the calibrated data and then plotting those metrics.
    - how to obtain non-wear windows and visualize them.
    - how to obtain sleep windows and visualize them.
-      - how we can filter sleep windows that overlap with non-wear periods.
 
 
-Example 1: Running the default processor
-----------------------------------------
+
+
+
+## Example 1: Running the default processor
+
 
 The `orchestrator` module of wristpy contains the default processor that will run the entire wristpy processing pipeline. This can be called as simply as:
 
@@ -28,7 +30,8 @@ results = orchestrator.run(
    output = 'path/to/save/file_name.csv'
 )
 ```
-This runs the processing pipeline with all the default arguments, creates an output `.csv` file, and will create a `results` object that contains the various output metrics (namely, enmo, angle-z, physical activity values, non-wear detection, sleep detection).
+This runs the processing pipeline with all the default arguments, creates an output `.csv` file, a `.json` file with the pipeline configuration parameters, and will create a `results` object that contains the various output metrics (namely; the specified physical activity metric, angle-z, physical activity classification values, non-wear status, and sleep status).
+
 
 The orchestrator can also process entire directories. The call to the orchestrator remains largely the same but now output is expected to be a directory and the desired filetype for the saved files **must** be specified:
 
@@ -46,10 +49,10 @@ results = orchestrator.run(
 
 We can visualize some of the outputs within the `results` object, directly, with the following scripts:
 
-Plot the ENMO across the entire data set:
+Plot the default physical activity metrics (ENMO) across the entire data set:
 ```python
 from matplotlib import pyplot as plt
-plt.plot(results.enmo.time, results.enmo.measurements)
+plt.plot(results.physical_activity_metric.time, results.physical_activity_metric.measurements)
 ```
 
 ![Example of the ENMO result](enmo_example1.png)
@@ -59,7 +62,7 @@ Plot the sleep windows with normalized angle-z data:
 from matplotlib import pyplot as plt
 
 plt.plot(results.anglez.time, results.anglez.measurements/90)
-plt.plot(results.sleep_windows_epoch.time, results.sleep_windows_epoch.measurements)
+plt.plot(results.sleep_status.time, results.sleep_status.measurements)
 plt.legend(['Angle Z', 'Sleep Windows'])
 plt.show()
 ```
@@ -70,8 +73,15 @@ We can also view and process these outputs from the saved `.csv` output file:
 ```python
 import polars as pl
 import matplotlib.pyplot as plt
-output_results = pl.read_csv('path/to/save/file_name.csv', try_parse_dates=True)
-phys_activity = output_results['physical_activity_levels'].cast(pl.Categorical).to_physical()
+
+activity_mapping = {
+    "inactive": 0,
+    "light": 1,
+    "moderate": 2,
+    "vigorous": 3
[…]
