### 2025-06-26 #8224 @angerhang
> 👋 [@philerooski](https://github.com/philerooski) & [@matt002](https://github.com/matt002) & [@arianesasso](https://github.com/arianesasso) & [@angerhang](https://github.com/angerhang) -- would you be available and willing to review this submission for JOSS?

Hey @sappelhoff thanks for the invite. Great to see some Jacobs alumni in this community haha. I'd really love to but still have 3-4 papers to review at the moment so have to politely decline this request.  Potentially my colleagues @chanshing and @aidanacquah could be great candidates to review this.
