### 2025-06-26 #8224 @sappelhoff
@angerhang what a coincidence, I did not even realize this was you until you mentioned it. Great to see you 👋 and thanks for your response and the referrals. Good luck with your reviews.

@chanshing @aidanacquah please let me know whether you would be willing to review this submission for JOSS.
