+++ src/wristpy/processing/analytics.py
@@ -38,7 +38,9 @@ class AbstractSleepDetector(abc.ABC):
         pass
 
     @abc.abstractmethod
-    def run_sleep_detection(self) -> List[SleepWindow]:
+    def run_sleep_detection(
+        self,
+    ) -> Tuple[List[SleepWindow], models.Measurement, models.Measurement]:
         """Sleep Detector must contain a run_sleep_detection function.
 
         The function must return a list of SleepWindow objects.
@@ -68,7 +70,9 @@ class GgirSleepDetection(AbstractSleepDetector):
         """
         self.anglez = anglez
 
-    def run_sleep_detection(self) -> List[SleepWindow]:
+    def run_sleep_detection(
+        self,
+    ) -> Tuple[List[SleepWindow], models.Measurement, models.Measurement]:
         """Run the GGIR sleep detection.
 
         This algorithm uses the angle-z data to first find potential sleep periods
@@ -91,7 +95,7 @@ class GgirSleepDetection(AbstractSleepDetector):
         logger.debug(
             "Sleep detection complete. Windows detected: %s", len(sleep_onset_wakeup)
         )
-        return sleep_onset_wakeup
+        return [sleep_onset_wakeup, spt_window, sib_periods]
 
     def _spt_window(
         self, anglez_data: models.Measurement, threshold: float = 0.2
