### 2025-06-27 7bbbc08
Move cleanup to own function in analytics
