### 2025-07-01 7c8ca6c
Bump the production-dependencies group across 1 directory with 4 updates
Bumps the production-dependencies group with 4 updates in the / directory: [polars](https://github.com/pola-rs/polars), [pyarrow](https://github.com/apache/arrow), [pydantic-settings](https://github.com/pydantic/pydantic-settings) and [typer](https://github.com/fastapi/typer).


Updates `polars` from 1.30.0 to 1.31.0
- [Release notes](https://github.com/pola-rs/polars/releases)
- [Commits](https://github.com/pola-rs/polars/compare/py-1.30.0...py-1.31.0)

Updates `pyarrow` from 19.0.1 to 20.0.0
- [Release notes](https://gi
[…]
