### 2025-07-02 2e774fa
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
