### 2025-07-02 8aae551
custom exception
