### 2025-07-03 32eea81
Update wristpy_tutorial.md
This line is missing from the tutorial
