### 2025-07-03 840399a
doc string
