### 2025-07-03 d7cff6b
Feature/Issue 212/Feedback (#230)
* default file type, supressing traceback

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* test errors

* test error.

* custom exception

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
