### 2025-07-03 51972f0
modifying measure inplace
