+++ src/wristpy/processing/mims.py
@@ -687,12 +687,9 @@ def aggregate_mims(
 
     if truncate:
         truncate_threshold = 1e-4 * epoch * sampling_rate
-        aggregated_measure.measurements = np.where(
-            aggregated_measure.measurements <= truncate_threshold,
-            0,
-            aggregated_measure.measurements,
-        )
-
+        aggregated_measure.measurements[
+            aggregated_measure.measurements <= truncate_threshold
+        ] = 0
     return aggregated_measure
 
 
