### 2025-07-03 d83646e
truncation fix (#226)
* truncation fix

* testing

* testing

* test fix

* mark down

* review changes.

* doc string

* modifying measure inplace
