### 2025-07-04 08a7536
Changes for review
Remove abstract sleepdetector class as we have no plans to extend to new sleep detection algorithms in the immediate future
Add new sleep parameters data class to store internal sleep parameters to be used in post-proc
