### 2025-07-04 eac234e
Remove tuple returns and run function twice
