### 2025-07-04 6bc6a88
Update test_analytics.py
No need for List in 3.10+
