+++ tests/unit/test_analytics.py
@@ -2,7 +2,6 @@
 
 import datetime
 import math
-from typing import List
 
 import numpy as np
 import polars as pl
