+++ src/wristpy/core/orchestrator.py
@@ -20,17 +20,24 @@ logger = config.get_logger()
 
 VALID_FILE_TYPES = (".csv", ".parquet")
 
+DEFAULT_ACTIVITY_THRESHOLDS = {
+    "enmo": (0.0563, 0.1916, 0.6958),
+    "mad": (0.029, 0.338, 0.604),
+    "ag_count": (100, 3000, 5200),
+    "mims": (10.558, 15.047, 19.614),
+}
+
 
 def run(
     input: Union[pathlib.Path, str],
     output: Optional[Union[pathlib.Path, str]] = None,
-    thresholds: Optional[Tuple[float, float, float]] = None,
+    thresholds: Optional[Sequence[Tuple[float, float, float]]] = None,
     calibrator: Union[
         None,
         Literal["ggir", "gradient"],
     ] = "gradient",
     epoch_length: float = 5,
-    activity_metric: Literal["enmo", "mad", "ag_count", "mims"] = "enmo",
+    activity_metric: Sequence[Literal["enmo", "mad", "ag_count", "mims"]] = ("enmo",),
     nonwear_algorithm: Sequence[Literal["ggir", "cta", "detach"]] = ["ggir"],
     verbosity: int = logging.WARNING,
     output_filetype: Literal[".csv", ".parquet"] = ".csv",
@@ -89,19 +96,27 @@ def run(
     input = pathlib.Path(input)
     output = pathlib.Path(output) if output is not None else None
 
-    if activity_metric == "enmo":
-        thresholds = thresholds or (0.0563, 0.1916, 0.6958)
-    elif activity_metric == "mad":
-        thresholds = thresholds or (0.029, 0.338, 0.604)
-    elif activity_metric == "ag_count":
-        thresholds = thresholds or (100, 3000, 5200)
-    elif activity_metric == "mims":
-        thresholds = thresholds or (10.558, 15.047, 19.614)
+    if thresholds is not None:
+        if len(activity_metric) != len(thresholds):
+            raise ValueError(
+                "Number of thresholds did not match the number of activity metrics. "
+                "Provide one threshold tuple per metric or use None for defaults."
+            )
+        metrics_dict = dict(zip(activity_metric, thresholds))
 
-    if not (0 <= thresholds[0] < thresholds[1] < thresholds[2]):  # type: ignore
-        message = "Threshold values must be >=0, unique, and in ascending order."
-        logger.error(message)
-        raise ValueError(message)
+    else:
+        metrics_dict = {
+            metric: DEFAULT_ACTIVITY_THRESHOLDS[metric] for metric in activity_metric
+        }
+
+    for metric, thresh in metrics_dict.items():
+        if not (0 <= thresh[0] < thresh[1] < thresh[2]):
+            message = (
+                f"Invalid thresholds for {metric}."
+                f" Threshold values must be >=0, unique, and in ascending order."
+            )
+            logger.error(message)
+            raise ValueError(message)
 
     if input.is_file():
         return _run_file(
@@ -131,7 +146,7 @@ def run(
 def _run_directory(
     input: pathlib.Path,
     output: Optional[pathlib.Path] = None,
-    thresholds: Tuple[float, float, float] = (0.0563, 0.1916, 0.6958),
+    thresholds: Optional[Sequence[Tuple[float, float, float]]] = None,
     calibrator: Union[
         None,
         Literal["ggir", "gradient"],
@@ -140,7 +155,7 @@ def _run_directory(
     nonwear_algorithm: Sequence[Literal["ggir", "cta", "detach"]] = ["ggir"],
     verbosity: int = logging.WARNING,
     output_filetype: Literal[".csv", ".parquet"] = ".csv",
-    activity_metric: Literal["enmo", "mad", "ag_count", "mims"] = "enmo",
[…]
