### 2025-07-09 52927ca
loop over metric
