### 2025-07-18 7f53b22
revert bug check change
