+++ src/wristpy/core/orchestrator.py
@@ -472,8 +472,6 @@ def _compute_activity(
             name=activity_metric,
         )
 
-    enmo_return = metrics.euclidean_norm_minus_one(
+    return metrics.euclidean_norm_minus_one(
         acceleration, epoch_length=epoch_length, name=activity_metric
     )
-
-    return enmo_return
