### 2025-07-18 #8224 @sappelhoff
👋 @jsgalan, @djmitche, @luciorq would any of you be available to review this submission for JOSS?
