### 2025-07-22 9845e75
cli str fix.
