### 2025-07-23 #8224 @Trybnetic
@sappelhoff  Thank you very much for thinking of me as a potential reviewer. The package looks super interesting and I would love to review. Unfortunately, I am on sabbatical and will be traveling from tomorrow onward and can therefore not provide a timely review. If this issue is still open November/December, I am more than happy to contribute.

@claaslendt I could imagine this package might be also interesting for you?
