+++ tests/unit/test_cli.py
@@ -5,6 +5,7 @@ import pathlib
 
 import pytest
 import pytest_mock
+import typer
 from typer import testing
 
 from wristpy.core import cli, orchestrator
@@ -23,7 +24,6 @@ def test_main_default(
 ) -> None:
     """Test cli with only necessary arguments."""
     mock_run = mocker.patch.object(orchestrator, "run")
-    default_thresholds = None
 
     result = create_typer_cli_runner.invoke(cli.app, [str(sample_data_gt3x)])
 
@@ -31,9 +31,9 @@ def test_main_default(
     mock_run.assert_called_once_with(
         input=sample_data_gt3x,
         output=None,
-        thresholds=default_thresholds,
+        thresholds=None,
         calibrator=None,
-        activity_metric="enmo",
+        activity_metric=["enmo"],
         epoch_length=5,
         nonwear_algorithm=["ggir"],
         verbosity=logging.INFO,
@@ -41,69 +41,32 @@ def test_main_default(
     )
 
 
-def test_main_enmo_default(
+@pytest.mark.parametrize(
+    "metric",
+    [
+        "enmo",
+        "mad",
+        "ag_count",
+        "mims",
+    ],
+)
+def test_main_with_metrics(
     mocker: pytest_mock.MockerFixture,
     sample_data_gt3x: pathlib.Path,
     create_typer_cli_runner: testing.CliRunner,
-) -> None:
-    """Test that correct enmo default thresholds are pulled."""
-    mock_run = mocker.patch.object(orchestrator, "_run_file")
-    default_thresholds = (0.0563, 0.1916, 0.6958)
-
-    create_typer_cli_runner.invoke(cli.app, ([str(sample_data_gt3x), "-a", "enmo"]))
-
-    mock_run.assert_called_once_with(
-        input=sample_data_gt3x,
-        output=None,
-        thresholds=default_thresholds,
-        calibrator=None,
-        activity_metric="enmo",
-        epoch_length=5,
-        nonwear_algorithm=["ggir"],
-        verbosity=logging.INFO,
-    )
-
-
-def test_main_mad_default(
-    mocker: pytest_mock.MockerFixture,
-    sample_data_gt3x: pathlib.Path,
-    create_typer_cli_runner: testing.CliRunner,
-) -> None:
-    """Test that correct mad default thresholds are pulled."""
-    mock_run = mocker.patch.object(orchestrator, "_run_file")
-    default_thresholds = (0.029, 0.338, 0.604)
-
-    create_typer_cli_runner.invoke(cli.app, ([str(sample_data_gt3x), "-a", "mad"]))
-
-    mock_run.assert_called_once_with(
-        input=sample_data_gt3x,
-        output=None,
-        thresholds=default_thresholds,
-        calibrator=None,
-        activity_metric="mad",
-        nonwear_algorithm=["ggir"],
-        epoch_length=5,
-        verbosity=logging.INFO,
-    )
-
-
-def test_main_agcount_default(
-    mocker: pytest_mock.MockerFixture,
-    sample_data_gt3x: pathlib.Path,
-    create_typer_cli_runner: testing.CliRunner,
+    metric: str,
 ) -> None:
     """Test that correct ag_count default thresholds are pulled."""
     mock_run = mocker.patch.object(orchestrator, "_run_file")
-    default_thresholds = (100, 3000, 5200)
 
-    create_typer_cli_runner.invoke(cli.app, ([str(sample_data_gt3x), "-a", "ag_count"]))
+    create_typer_cli_runner.invoke(cli.app, ([str(sample_data_gt3x), "-a", metric]))
 
     mock_run.assert_called_once_with(
         input=sample_data_gt3x,
         output=None,
-        thresholds=default_thresholds,
[…]
