### 2025-07-23 4a5fe1f
Feature/issue 177/output sib spt windows (#225)
* Initial changes to return SIB/SPT

* More small changes

* output sib/spt

Use the nonwear_utils package to fix timing for sib/spt
Update orchestrator +tests to output data

* Move cleanup to own function in analytics

* Renaming utils module

* Update documentation

* Update wristpy_tutorial.md

This line is missing from the tutorial

* Changes for review

Remove abstract sleepdetector class as we have no plans to extend to new sleep detection algorithms in the immediate future
Add new sleep parameters data class to store internal sleep parame
[…]
