### 2025-07-25 c9ce6ef
post-merge fixes.
