### 2025-07-28 4b29e5a
multiple metrics tests for cli and orchestrator.
