### 2025-07-28 c4a7e1e
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/python-poetry/poetry: 1.7.0 → 2.1.3](https://github.com/python-poetry/poetry/compare/1.7.0...2.1.3)
- [github.com/astral-sh/ruff-pre-commit: v0.4.8 → v0.12.5](https://github.com/astral-sh/ruff-pre-commit/compare/v0.4.8...v0.12.5)
- [github.com/pre-commit/mirrors-mypy: v1.8.0 → v1.17.0](https://github.com/pre-commit/mirrors-mypy/compare/v1.8.0...v1.17.0)
- [github.com/macisamuele/language-formatters-pre-commit-hooks: v2.12.0 → v2.15.0](https://github.com/macisamuele/language-formatters-pre-commit-hooks/compare/v2.12.0...v2.15.0)
- [gi
[…]
