+++ .pre-commit-config.yaml
@@ -2,23 +2,23 @@ fail_fast: false
 
 repos:
 - repo: https://github.com/python-poetry/poetry
-  rev: 1.7.0
+  rev: 2.1.3
   hooks:
   - id: poetry-check
 
 - repo: https://github.com/astral-sh/ruff-pre-commit
-  rev: v0.4.8
+  rev: v0.12.5
   hooks:
   - id: ruff
   - id: ruff-format
 
 - repo: https://github.com/pre-commit/mirrors-mypy
-  rev: v1.8.0
+  rev: v1.17.0
   hooks:
   - id: mypy
     args: [--ignore-missing-imports]
 - repo: https://github.com/macisamuele/language-formatters-pre-commit-hooks
-  rev: v2.12.0
+  rev: v2.15.0
   hooks:
   - id: pretty-format-yaml
     args:
@@ -32,7 +32,7 @@ repos:
     - --no-sort
 
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v4.5.0
+  rev: v5.0.0
   hooks:
   - id: check-case-conflict
   - id: end-of-file-fixer
