+++ src/wristpy/core/cli.py
@@ -88,7 +88,7 @@ def _parse_thresholds(thresholds: list[str]) -> list[tuple[float, float, float]]
             values = [float(part) for part in parts]
         except ValueError:
             raise typer.BadParameter(f"Invalid float in threshold: {triplet_str}")
-        parsed.append(tuple(values))
+        parsed.append((values[0], values[1], values[2]))
     return parsed
 
 
