### 2025-08-01 89db955
Feature/issue 166/multiple metrics (#233)
* orchestrator input changes

* loop over metric

* names to function calls.

* multipule metrics, pal

* multiple metrics, pal working fix for orchestrator

* cli fix

* revert bug check change

* threshold change

* threshold parsing.

* cli str fix.

* working tests

* documentation.

* post-merge fixes.

* Json error

* ruff error

* reverting old test change.

* typer error

* mypy errors

* test for typer error.

* removing validation.

* removing redundant test causing error.

* multiple metrics tests for cli and orchestrator.

* extra test for 
[…]
