+++ src/wristpy/core/orchestrator.py
@@ -20,17 +20,24 @@ logger = config.get_logger()
 
 VALID_FILE_TYPES = (".csv", ".parquet")
 
+DEFAULT_ACTIVITY_THRESHOLDS = {
+    "enmo": (0.0563, 0.1916, 0.6958),
+    "mad": (0.029, 0.338, 0.604),
+    "ag_count": (100, 3000, 5200),
+    "mims": (10.558, 15.047, 19.614),
+}
+
 
 def run(
     input: Union[pathlib.Path, str],
     output: Optional[Union[pathlib.Path, str]] = None,
-    thresholds: Optional[Tuple[float, float, float]] = None,
+    thresholds: Optional[Sequence[Tuple[float, float, float]]] = None,
     calibrator: Union[
         None,
         Literal["ggir", "gradient"],
     ] = "gradient",
     epoch_length: float = 5,
-    activity_metric: Literal["enmo", "mad", "ag_count", "mims"] = "enmo",
+    activity_metric: Sequence[Literal["enmo", "mad", "ag_count", "mims"]] = ["enmo"],
     nonwear_algorithm: Sequence[Literal["ggir", "cta", "detach"]] = ["ggir"],
     verbosity: int = logging.WARNING,
     output_filetype: Literal[".csv", ".parquet"] = ".csv",
@@ -44,19 +51,22 @@ def run(
     Output file names will be derived from original file names in the case of directory
     processing.
 
-
     Args:
         input: Path to the input file or directory of files to be read. Currently,
             this supports .bin and .gt3x
         output: Path to directory data will be saved to. If processing a single file the
             path should end in the save file name in either .csv or .parquet formats.
         thresholds: The cut points for the light, moderate, and vigorous thresholds,
-            given in that order. Values must be asscending, unique, and greater than 0.
-            Default values are optimized for subjects ages 7-11 [1][3].
+            given in that order. One threshold tuple must be provided for each activity
+            metric, in the same order the metrics were specified. To use default values
+            for all metrics, leave thresholds as None. Values must be ascending, unique,
+            and greater than 0. Default values are optimized for subjects ages 7-11 [1]
+            [3].
         calibrator: The calibrator to be used on the input data.
         epoch_length: The temporal resolution in seconds, the data will be down sampled
             to. It must be > 0.0.
-        activity_metric: The metric to be used for physical activity categorization.
+        activity_metric: The metric(s) to be used for physical activity categorization.
+            Multiple metrics can be specified as a sequence.
         nonwear_algorithm: The algorithm to be used for nonwear detection.
         verbosity: The logging level for the logger.
         output_filetype: Specifies the data format for the save files. Only used when
@@ -67,9 +77,10 @@ def run(
         dictionary of OrchestratorResults objects.
 
     Raises:
+        ValueError: If the number of physical activity thresholds does not match the
+            number of provided activity metrics.
         ValueError: If the physical activity thresholds are not unique or not in
             ascending order.
-
 
     References:
         [1] Hildebrand, M., et al. (2014). Age group comparability of raw accelerometer
@@ -89,19 +100,26 @@ def run(
     input = pathlib.Path(input)
     output = pathlib.Path(output) if output is not None else None
 
-    if activity_metric == "enmo":
-        thresholds = thresholds or (0.0563, 0.1916, 0.6958)
-    elif activity_metric == "mad":
-        thresholds = thresholds or (0.029, 0.338, 0.604)
-    elif activity_metric == "ag_count":
-        thresholds = thresholds or (100, 3000, 5200)
-    elif activity_metric == "mims":
-        thresholds = thresholds or (10.558, 15.047, 19.614)
+    if thresholds is not None:
+        if len(activity_metric) != len(thresholds):
+            raise ValueError(
[…]
