### 2025-08-11 childmindresearch/wristpy#236 @Asanto32
Task: [JOSS paper] Fix included documentation
### Description

Currently the include statements (below) that are in the `__init__.py` do not actually work, and the documentation cannot be run from the terminal `print(wristpy.__doc__)`
```
""".. include:: ../../README.md
.. include:: ../../docs/wristpy_tutorial.md
"""
```

See details here: https://github.com/openjournals/joss-reviews/issues/8637#issuecomment-3166241051

### Tasks

- [ ] Properly include readme and tutorial (a quick investigation seems like this is Sphinx/rts notation)

### Freeform Notes

_No response_
