+++ docs/source/api/wristpy.processing.metrics.rst
@@ -1,0 +1,7 @@
+metrics
+=================================
+
+.. automodule:: wristpy.processing.metrics
+   :members:
+   :undoc-members:
+   :show-inheritance:
