+++ docs/source/api/wristpy.core.exceptions.rst
@@ -1,0 +1,7 @@
+exceptions
+==============================
+
+.. automodule:: wristpy.core.exceptions
+   :members:
+   :undoc-members:
+   :show-inheritance:
