+++ docs/source/api/wristpy.io.rst
@@ -1,0 +1,16 @@
+IO Module
+==================
+
+.. automodule:: wristpy.io
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   wristpy.io.readers
+   wristpy.io.writers
