+++ docs/source/api/wristpy.io.writers.rst
@@ -1,0 +1,15 @@
+writers
+==========================
+
+.. automodule:: wristpy.io.writers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 4
+
+   wristpy.io.writers.writers
