+++ docs/source/api/wristpy.core.rst
@@ -1,0 +1,20 @@
+Core Module
+====================
+
+.. automodule:: wristpy.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. toctree::
+   :maxdepth: 2
+
+   wristpy.core.cli
+   wristpy.core.computations
+   wristpy.core.config
+   wristpy.core.exceptions
+   wristpy.core.models
+   wristpy.core.orchestrator
