+++ docs/source/api/wristpy.core.computations.rst
@@ -1,0 +1,7 @@
+computations
+================================
+
+.. automodule:: wristpy.core.computations
+   :members:
+   :undoc-members:
+   :show-inheritance:
