+++ docs/source/api/modules.rst
@@ -1,0 +1,7 @@
+wristpy
+=======
+
+.. toctree::
+   :maxdepth: 4
+
+   wristpy
