+++ docs/source/api/wristpy.io.writers.writers.rst
@@ -1,0 +1,7 @@
+writers
+=================================
+
+.. automodule:: wristpy.io.writers.writers
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
