+++ docs/source/api/wristpy.rst
@@ -1,17 +1,0 @@
-wristpy package
-===============
-
-.. automodule:: wristpy
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   wristpy.core
-   wristpy.io
-   wristpy.processing
