+++ docs/source/api/wristpy.io.writers.writers.rst
@@ -1,7 +1,13 @@
-writers
-=================================
+wristpy.io.writers.writers
+==========================
 
 .. automodule:: wristpy.io.writers.writers
-   :members:
-   :undoc-members:
-   :show-inheritance:
+
+   
+   .. rubric:: Classes
+
+   .. autosummary::
+   
+      Any
+      OrchestratorResults
+   
