+++ docs/source/api/wristpy.core.computations.rst
@@ -1,7 +1,15 @@
-computations
-================================
+wristpy.core.computations
+=========================
 
 .. automodule:: wristpy.core.computations
-   :members:
-   :undoc-members:
-   :show-inheritance:
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      moving_mean
+      moving_median
+      moving_std
+      resample
+   
