+++ docs/source/api/wristpy.processing.metrics.rst
@@ -1,7 +1,19 @@
-metrics
-=================================
+wristpy.processing.metrics
+==========================
 
 .. automodule:: wristpy.processing.metrics
-   :members:
-   :undoc-members:
-   :show-inheritance:
+
+   
+   .. rubric:: Functions
+
+   .. autosummary::
+   
+      actigraph_activity_counts
+      angle_relative_to_horizontal
+      combined_temp_accel_detect_nonwear
+      detach_nonwear
+      detect_nonwear
+      euclidean_norm_minus_one
+      mean_amplitude_deviation
+      monitor_independent_movement_summary_units
+   
