### 2025-08-13 581cbab
more api automation changes.
