+++ .github/workflows/docs.yaml
@@ -7,29 +7,34 @@ on:
   workflow_run:
     workflows: [Python Tests]
     types:
-    - completed
+      - completed
     branches:
-    - main
+      - main
 
 jobs:
   build-and-publish-docs:
     if: ${{ github.event.workflow_run.conclusion == 'success' }}
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
-    - name: Fetch LFS files
-      run: git lfs pull
-    - uses: actions/setup-python@v5
-      with:
-        python-version: '3.11'
-    - name: Install dependencies
-      run: |
-        python -m pip install poetry
-        poetry install --with docs
-    - name: Build docs
-      run: |
-        poetry run pdoc src/wristpy -o docs_build -t docs/pdoc-theme --docformat google
-        touch docs_build/.nojekyll
-    - uses: JamesIves/github-pages-deploy-action@v4
-      with:
-        folder: docs_build
+      - uses: actions/checkout@v4
+
+      - name: Fetch LFS files
+        run: git lfs pull
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: '3.11'
+
+      - name: Install dependencies
+        run: |
+          python -m pip install poetry
+          poetry install --with docs
+
+      - name: Build Sphinx docs
+        run: |
+          poetry run sphinx-build -b html docs/source docs/_build/html
+          touch docs/_build/html/.nojekyll
+
+      - uses: JamesIves/github-pages-deploy-action@v4
+        with:
+          folder: docs/_build/html
