+++ src/wristpy/__init__.py
@@ -1,3 +1,5 @@
 """.. include:: ../../README.md
 .. include:: ../../docs/wristpy_tutorial.md
 """  # noqa: D205, D415
+
+__doc__ = """https://childmindresearch.github.io/wristpy/wristpy.html"""
