### 2025-08-13 f4100d3
command change and duplicate code.
