+++ .github/workflows/docs.yaml
@@ -39,4 +39,7 @@ jobs:
 
       - uses: JamesIves/github-pages-deploy-action@v4
         with:
-          folder: docs/build/html  
+          branch: gh-pages         
+          folder: docs/build/html   
+          clean: true               
+ 
