+++ _sources/development.rst.txt
@@ -1,0 +1,9 @@
+Development Guide
+=================
+
+Project Governance
+------------------
+
+- `Contributing Guidelines <https://github.com/childmindresearch/wristpy/blob/main/CONTRIBUTING.md>`_ - How to contribute to wristpy
+- `Security Policy <https://github.com/childmindresearch/wristpy/blob/main/SECURITY.md>`_ - How to report security issues  
+- `Code of Conduct <https://github.com/childmindresearch/wristpy/blob/main/CODE_OF_CONDUCT.md>`_ - Community standards
