+++ _sources/api/wristpy.core.rst.txt
@@ -1,0 +1,21 @@
+core
+============
+
+.. automodule:: wristpy.core
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Submodules
+----------
+
+.. autosummary::
+   :toctree: .
+   :recursive:
+
+   cli
+   computations
+   config
+   exceptions
+   models
+   orchestrator
