+++ src/wristpy/__init__.py
@@ -1,5 +1,5 @@
-""".. include:: ../../README.md
-.. include:: ../../docs/wristpy_tutorial.md
-"""  # noqa: D205, D415
+"""This is wristpy, a Python package for wrist-worn accelerometer data processing."""
 
-__doc__ = """https://childmindresearch.github.io/wristpy/wristpy.html"""
+__doc__ = (
+    " For full documentation, please see: https://childmindresearch.github.io/wristpy/"
+)
