+++ .github/workflows/docs.yaml
@@ -33,7 +33,7 @@ jobs:
         run: |
           poetry run make html
           touch build/html/.nojekyll
-          ls -la build/html | sed 's/^/BUILD: /'   # sanity check in logs
+          ls -la build/html | sed 's/^/BUILD: /'   
 
       - name: Deploy to gh-pages
         uses: JamesIves/github-pages-deploy-action@v4
