+++ .github/workflows/docs.yaml
@@ -4,15 +4,16 @@ permissions:
   contents: write
 
 on:
-  push:
-    branches:
-      - feature/issue-218/sphinx
-      - main
-  workflow_dispatch:
+  workflow_run:
+    workflows: [Python Tests]
+    types: [completed]
+    branches: [main]
 
 jobs:
   build-and-publish-docs:
+    if: ${{ github.event.workflow_run.conclusion == 'success' }}
     runs-on: ubuntu-latest
+
     steps:
       - uses: actions/checkout@v4
 
@@ -31,9 +32,10 @@ jobs:
       - name: Build Sphinx docs
         working-directory: docs
         run: |
-          poetry run make html
+          # Fail on warnings to catch broken refs; drop -W if you prefer lenient builds
+          poetry run make html SPHINXOPTS="-W"
           touch build/html/.nojekyll
-          ls -la build/html | sed 's/^/BUILD: /'   
+          ls -la build/html | sed 's/^/BUILD: /'
 
       - name: Deploy to gh-pages
         uses: JamesIves/github-pages-deploy-action@v4
