+++ docs/source/conf.py
@@ -9,7 +9,7 @@ sys.path.insert(0, str(src_path))
 
 project = "wristpy"
 copyright = "2025, Child Mind Institute"
-author = """ Adam Santorelli, Florian Rupprecht, Freymon Perez, Greg Kiar, 
+author = """ Adam Santorelli, Florian Rupprecht, Freymon Perez, Greg Kiar,
 Reinder Vos de Wael"""
 
 release = "0.2.0"
