+++ docs/source/api/wristpy.core.config.rst
@@ -3,11 +3,11 @@
 
 .. automodule:: wristpy.core.config
 
-   
+
    .. rubric:: Functions
 
    .. autosummary::
-   
+
       get_logger
       get_version
-   
+
