+++ docs/source/api/wristpy.io.writers.writers.rst
@@ -3,11 +3,11 @@ wristpy.io.writers.writers
 
 .. automodule:: wristpy.io.writers.writers
 
-   
+
    .. rubric:: Classes
 
    .. autosummary::
-   
+
       Any
       OrchestratorResults
-   
+
