+++ docs/source/api/wristpy.processing.analytics.rst
@@ -3,21 +3,21 @@
 
 .. automodule:: wristpy.processing.analytics
 
-   
+
    .. rubric:: Functions
 
    .. autosummary::
-   
+
       compute_physical_activty_categories
       dataclass
       sleep_bouts_cleanup
       sleep_cleanup
-   
+
    .. rubric:: Classes
 
    .. autosummary::
-   
+
       GgirSleepDetection
       SleepParameters
       SleepWindow
-   
+
