+++ docs/source/api/wristpy.io.readers.readers.rst
@@ -3,11 +3,11 @@ wristpy.io.readers.readers
 
 .. automodule:: wristpy.io.readers.readers
 
-   
+
    .. rubric:: Functions
 
    .. autosummary::
-   
+
       read_watch_data
       unix_epoch_time_to_polars_datetime
-   
+
