+++ pyproject.toml
@@ -35,7 +35,6 @@ ruff = ">=0.7.1,<0.9.0"
 pytest-mock = "^3.14.0"
 sphinx-autobuild = "^2024.2.4"
 
-
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.0.0"
 furo = "^2024.1.0"
