+++ .github/workflows/docs.yaml
@@ -4,16 +4,15 @@ permissions:
   contents: write
 
 on:
-  workflow_run:
-    workflows: [Python Tests]
-    types: [completed]
-    branches: [main]
+  push:
+    branches:
+      - feature/issue-218/sphinx
+      - main
+  workflow_dispatch:
 
 jobs:
   build-and-publish-docs:
-    if: ${{ github.event.workflow_run.conclusion == 'success' }}
     runs-on: ubuntu-latest
-
     steps:
       - uses: actions/checkout@v4
 
@@ -31,11 +30,11 @@ jobs:
 
       - name: Build Sphinx docs
         working-directory: docs
+
         run: |
-          # Fail on warnings to catch broken refs; drop -W if you prefer lenient builds
-          poetry run make html SPHINXOPTS="-W"
+          poetry run make html
           touch build/html/.nojekyll
-          ls -la build/html | sed 's/^/BUILD: /'
+          ls -la build/html | sed 's/^/BUILD: /'   
 
       - name: Deploy to gh-pages
         uses: JamesIves/github-pages-deploy-action@v4
