### 2025-08-22 efc1272
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
