+++ docs/source/api/wristpy.io.writers.writers.rst
@@ -10,4 +10,3 @@ wristpy.io.writers.writers
 
       Any
       OrchestratorResults
-
