+++ docs/source/api/wristpy.core.orchestrator.rst
@@ -9,4 +9,3 @@
    .. autosummary::
 
       run
-
