+++ docs/source/api/wristpy.core.config.rst
@@ -10,4 +10,3 @@
 
       get_logger
       get_version
-
