+++ docs/source/api/wristpy.core.models.rst
@@ -17,4 +17,3 @@
       BaseModel
       Measurement
       WatchData
-
