+++ docs/source/api/wristpy.core.exceptions.rst
@@ -14,4 +14,3 @@
       LoggedException
       NoMotionError
       SphereCriteriaError
-
