### 2025-08-22 b21b41e
ruff lint ignore.
