+++ pyproject.toml
@@ -106,6 +106,7 @@ convention = "google"
 
 [tool.ruff.per-file-ignores]
 "tests/**/*.py" = []
+"docs/source/conf.py" = ["INP001"]
 
 [build-system]
 requires = ["poetry-core>=1.2.0"]
