+++ pyproject.toml
@@ -55,6 +55,7 @@ testpaths = [
 
 [tool.mypy]
 ignore_missing_imports = true
+exclude = ["docs/"]
 
 [tool.ruff]
 exclude = [
