+++ .github/workflows/docs.yaml
@@ -15,27 +15,21 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
-
     - name: Fetch LFS files
       run: git lfs pull
-
     - uses: actions/setup-python@v5
       with:
         python-version: '3.11'
-
     - name: Install dependencies
       run: |
         python -m pip install poetry
         poetry install --with docs
-
       - name: Build Sphinx docs
         working-directory: docs
-
         run: |
           poetry run make html
           touch build/html/.nojekyll
           ls -la build/html | sed 's/^/BUILD: /'
-
     - name: Deploy to gh-pages
       uses: JamesIves/github-pages-deploy-action@v4
       with:
