+++ .github/workflows/docs.yaml
@@ -24,7 +24,7 @@ jobs:
       run: |
         python -m pip install poetry
         poetry install --with docs
-      - name: Build Sphinx docs
+    - name: Build Sphinx docs
         working-directory: docs
         run: |
           poetry run make html
