### 2025-08-22 f840500
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
