### 2025-08-22 597175c
Update metrics.py
Should fix the failing tests due to new polars version update
