### 2025-09-09 7d1ac8e
Bump the development-dependencies group across 1 directory with 6 updates
Bumps the development-dependencies group with 6 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [pytest](https://github.com/pytest-dev/pytest) | `8.4.0` | `8.4.2` |
| [mypy](https://github.com/python/mypy) | `1.16.0` | `1.17.1` |
| [pre-commit](https://github.com/pre-commit/pre-commit) | `4.2.0` | `4.3.0` |
| [pytest-cov](https://github.com/pytest-dev/pytest-cov) | `6.2.1` | `7.0.0` |
| [ruff](https://github.com/astral-sh/ruff) | `0.8.6` | `0.12.12` |
| [pytest-mock](https://github.com/pytest-d
[…]
