### 2025-09-09 4c68cc5
Update dependency, adding permissions
