+++ .github/workflows/test.yaml
@@ -6,9 +6,14 @@ on:
     - main
   pull_request:
 
+  
 jobs:
   unit:
     runs-on: ${{ matrix.os }}
+      permissions:
+      contents: read
+      checks: write
+      pull-requests: write
     strategy:
       fail-fast: false
       matrix:
@@ -55,6 +60,8 @@ jobs:
 
   ruff:
     runs-on: ubuntu-latest
+    permissions:
+      contents: read
     steps:
     - uses: actions/checkout@v4
     - name: Install poetry
@@ -73,6 +80,8 @@ jobs:
 
   mypy:
     runs-on: ubuntu-latest
+    permissions:
+      contents: read
     steps:
     - uses: actions/checkout@v4
     - name: Install poetry
