+++ pyproject.toml
@@ -14,7 +14,7 @@ readme = "README.md"
 packages = [{include = "wristpy", from = "src"}]
 
 [tool.poetry.dependencies]
-python = ">=3.10, <3.13"
+python = ">=3.10, <3.14"
 polars = "^1.21.0"
 scikit-learn = "^1.4.1.post1"
 pyarrow = ">=18,<20"
