### 2025-09-09 0e72784
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
