+++ .github/workflows/test.yaml
@@ -6,7 +6,6 @@ on:
     - main
   pull_request:
 
-  
 jobs:
   unit:
     runs-on: ${{ matrix.os }}
