+++ .github/workflows/test.yaml
@@ -1,5 +1,8 @@
 name: Python Tests
 
+permissions:
+  contents: read
+  
 on:
   push:
     branches:
@@ -48,7 +51,7 @@ jobs:
           --verbose
       shell: bash
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v5
+      uses: codecov/codecov-action@5a1091511ad55cbe89839c7260b706298ca349f7
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         verbose: true
