+++ .github/workflows/pypi.yaml
@@ -41,7 +41,7 @@ jobs:
         poetry build
 
     - name: Publish to PyPi.
-      uses: pypa/gh-action-pypi-publish@release/v1
+      uses: pypa/gh-action-pypi-publish@ed0c53931b1dc9bd32cbe73a98c7f6766f8a527e
       with:
         user: __token__
         password: ${{ secrets.PYPI_KEY }}
