+++ .github/workflows/test.yaml
@@ -10,7 +10,7 @@ on:
 jobs:
   unit:
     runs-on: ${{ matrix.os }}
-      permissions:
+    permissions:
       contents: read
       checks: write
       pull-requests: write
@@ -53,7 +53,7 @@ jobs:
           --verbose
       shell: bash
     - name: Upload coverage to Codecov
-      uses: codecov/codecov-action@v5
+      uses: codecov/codecov-action@5a1091511ad55cbe89839c7260b706298ca349f7
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         verbose: true
