### 2025-09-09 ee4a366
Update pyproject.toml
bump min version of scipy
