### 2025-09-09 fb5aebe
Update test.yaml
Try moving this up
