+++ .github/workflows/test.yaml
@@ -24,15 +24,15 @@ jobs:
     - uses: actions/checkout@v4
     - name: Install poetry
       run: pipx install poetry
-    - uses: actions/setup-python@v5
-      with:
-        python-version: ${{ matrix.python_version }}
-        cache: pip
 
     - name: Install dependencies on macOS
       if: runner.os == 'macOS'
       run: |
         brew install libomp gfortran
+    - uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python_version }}
+        cache: pip
 
     - name: Install dependencies
       run: |
