### 2025-09-09 9567f03
Update pyproject.toml
Move to numpy > 2.1.0?
