+++ pyproject.toml
@@ -19,7 +19,7 @@ polars = "^1.21.0"
 scikit-learn = "^1.4.1.post1"
 pyarrow = ">=18,<20"
 actfast = "^1.1.2"
-numpy = ">=1.26.0, <3.0.0"
+numpy = ">=2.1.0, <3.0.0"
 pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.14.1"
