+++ _sources/api/wristpy.io.readers.readers.rst.txt
@@ -1,13 +1,0 @@
-wristpy.io.readers.readers
-==========================
-
-.. automodule:: wristpy.io.readers.readers
-
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      read_watch_data
-      unix_epoch_time_to_polars_datetime
-   
