+++ _sources/api/wristpy.io.rst.txt
@@ -1,17 +1,0 @@
-io
-==========
-
-.. automodule:: wristpy.io
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-.. autosummary::
-   :toctree: .
-   :recursive:
-
-   readers
-   writers
