+++ _sources/api/wristpy.core.cli.rst.txt
@@ -1,23 +1,0 @@
-﻿wristpy.core.cli
-================
-
-.. automodule:: wristpy.core.cli
-
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      main
-      version_check
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      ActivityMetric
-      Calibrator
-      Enum
-      NonwearAlgorithms
-      OutputFileType
-   
