+++ _sources/api/wristpy.core.computations.rst.txt
@@ -1,15 +1,0 @@
-﻿wristpy.core.computations
-=========================
-
-.. automodule:: wristpy.core.computations
-
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      moving_mean
-      moving_median
-      moving_std
-      resample
-   
