+++ _sources/api/wristpy.processing.idle_sleep_mode_imputation.rst.txt
@@ -1,12 +1,0 @@
-﻿wristpy.processing.idle\_sleep\_mode\_imputation
-================================================
-
-.. automodule:: wristpy.processing.idle_sleep_mode_imputation
-
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      impute_idle_sleep_mode_gaps
-   
