+++ _sources/api/wristpy.processing.rst.txt
@@ -1,21 +1,0 @@
-processing
-==================
-
-.. automodule:: wristpy.processing
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-.. autosummary::
-   :toctree: .
-   :recursive:
-
-   analytics
-   calibration
-   idle_sleep_mode_imputation
-   metrics
-   mims
-   processing_utils
