+++ _sources/api/wristpy.core.config.rst.txt
@@ -1,13 +1,0 @@
-﻿wristpy.core.config
-===================
-
-.. automodule:: wristpy.core.config
-
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      get_logger
-      get_version
-   
