+++ _sources/api/wristpy.processing.analytics.rst.txt
@@ -1,23 +1,0 @@
-﻿wristpy.processing.analytics
-============================
-
-.. automodule:: wristpy.processing.analytics
-
-   
-   .. rubric:: Functions
-
-   .. autosummary::
-   
-      compute_physical_activty_categories
-      dataclass
-      sleep_bouts_cleanup
-      sleep_cleanup
-   
-   .. rubric:: Classes
-
-   .. autosummary::
-   
-      GgirSleepDetection
-      SleepParameters
-      SleepWindow
-   
