### 2025-09-10 4d1e94c
Initial test of nimbledetach
failing mypy and ruff but just checking python tests cross platform
