+++ src/wristpy/processing/nimbaldetach.py
@@ -1,0 +1,304 @@
+"""Module for DETACH non-wear algorithm.
+
+This module is taken from Adam Vert https://github.com/nimbal/nimbaldetach
+and is licensed under the MIT License:
+
+MIT License
+
+Copyright (c) 2022 Adam Vert
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+from typing import Optional
+
+import numpy as np
+import pandas as pd
+from scipy import signal
+
+
+def nimbaldetach(
+    x_values: np.ndarray,
+    y_values: np.ndarray,
+    z_values: np.ndarray,
+    temperature_values: np.ndarray,
+    accel_freq: float = 75,
+    temperature_freq: float = 0.25,
+    std_thresh: float = 0.008,
+    low_temperature_cutoff: float = 26,
+    high_temperature_cutoff: float = 30,
+    temp_dec_roc: float = -0.2,
+    temp_inc_roc: float = 0.1,
+    num_axes: int = 2,
+    quiet: bool = False,
+) -> tuple[pd.DataFrame, np.ndarray]:
+    """Adam Vert's implementation of the DETACH algorithm.
+
+    Non-wear algorithm with a 5 minute minimum non-wear duration using absolute
+    temperature, temperature rate of change and accelerometer standard deviation of
+    3 axes to detect start and stop times of non-wear periods.
+
+    Args:
+        x_values: accelerometer x values
+        y_values: accelerometer y values
+        z_values: accelerometer z values
+        temperature_values: temperature values
+        accel_freq: frequency of accelerometer in Hz
+        temperature_freq: frequency of temperature sensor in Hz
+        std_thresh: the value which the std of an axis in the window must be below
+            to trigger non-wear.
+        low_temperature_cutoff: Low temperature for non-wear classification.
+        high_temperature_cutoff: High temperature cut off for wear classification.
+        temp_dec_roc: Decreasing temperature rate of change threshold for non-wear classification.
+        temp_inc_roc: Increasing temperature rate of change threshold for wear classification.
+        num_axes: the number of axes that must be below the std threshold to be considered nonwear.
+        quiet: Whether or not to quiet print statements
+
+    Returns:
+        A tuple with (start_stop_df, vert_nonwear_array) as defined below:
+
+        start_stop_df: A dataframe with the start and end datapoints of the non-wear bouts.
[…]
