+++ pyproject.toml
@@ -23,7 +23,6 @@ numpy = ">=2.1.0, <3.0.0"
 pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.14.1"
-scikit-digital-health = "^0.17.9"
 typer = "^0.15.2"
 
 [tool.poetry.group.dev.dependencies]
