### 2025-09-10 5e2fc2f
Fixing some ruff/mypy issues
