+++ src/wristpy/processing/nimbaldetach.py
@@ -46,7 +46,6 @@ def nimbaldetach(
     temp_dec_roc: float = -0.2,
     temp_inc_roc: float = 0.1,
     num_axes: int = 2,
-    quiet: bool = False,
 ) -> tuple[pd.DataFrame, np.ndarray]:
     """Adam Vert's implementation of the DETACH algorithm.
 
@@ -71,7 +70,7 @@ def nimbaldetach(
             wear classification.
         num_axes: The number of axes that must be below the std threshold
             to be considered nonwear.
-        quiet: Whether or not to quiet print statements.
+
 
     Returns:
         A tuple with (start_stop_df, vert_nonwear_array) as defined below:
@@ -80,25 +79,19 @@ def nimbaldetach(
         vert_nonwear_array: numpy array with length of the accelerometer data marked as
             either wear (0) or non-wear (1).
     """
-    if not quiet:
-        print("Starting DETACH Calculation...")
 
-    def filter_signal(
+    def lowpass_filter_signal(
         data: np.ndarray,
-        filt_type: str,
         sample_f: float,
         low_f: Optional[float] = None,
-        high_f: Optional[float] = None,
         filter_order: int = 2,
     ) -> np.ndarray:
-        """Function that fiters temperature data.
+        """Function that low pass fiters temperature data.
 
         Args:
             data: 1D numpy array of data to be filtered
-            filt_type: type of filter to be applied. One of 'lowpass', 'highpass', or 'bandpass'
-            sample_f: sampling frequency of data in Hz.
-            low_f: low frequency cutoff for filter in Hz.
-            high_f: high frequency cutoff for filter in Hz.
+            sample_f: Sampling rate, in Hz.
+            low_f: Low frequency cutoff for filter, in Hz.
             filter_order: order of the filter.
 
         Returns:
@@ -106,32 +99,20 @@ def nimbaldetach(
         """
         nyquist_freq = 0.5 * sample_f
         low = (low_f / nyquist_freq) if low_f is not None else None
-        high = (high_f / nyquist_freq) if high_f is not None else None
-        if filt_type == "lowpass":
-            wn = low
-        elif filt_type == "highpass":
-            wn = high
-        elif filt_type == "bandpass":
-            wn = [low, high]
-        b, a = signal.butter(N=filter_order, Wn=wn, btype=filt_type)
+        wn = low
+        b, a = signal.butter(N=filter_order, Wn=wn, btype="lowpass")
         filtered_data = signal.filtfilt(b, a, x=data)
         return filtered_data
 
-    vert_nonwear_array = np.zeros(
-        len(x_values)
-    )  # This array will contain 0 for wear and 1 for non-wear
+    vert_nonwear_array = np.zeros(len(x_values))
     vert_nonwear_start_datapoints = []
     vert_nonwear_end_datapoints = []
 
-    # Create seies of 1 minute rolling std
-    x_std_fwd = (
-        pd.Series(x_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
-    )  # 1 minute forward looking STD
+    x_std_fwd = pd.Series(x_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
     y_std_fwd = pd.Series(y_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
     z_std_fwd = pd.Series(z_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
-    x_std_back = (
-        pd.Series(x_values).rolling(round(accel_freq * 60)).std()
-    )  # 1 minute backward looking STD
+
+    x_std_back = pd.Series(x_values).rolling(round(accel_freq * 60)).std()
[…]
