### 2025-09-11 08d2594
Update nimbaldetach.py
Moved to private function outside nimbaldetach
