+++ .github/workflows/test.yaml
@@ -2,7 +2,7 @@ name: Python Tests
 
 permissions:
   contents: read
-  
+
 on:
   push:
     branches:
