### 2025-09-11 childmindresearch/wristpy#248 @Asanto32
Task: Remove skdh dependency and replace with direct implementation of `nimbaldetach`
### Description

skdh has been causing issues with 3.13 support across platforms, we will replace our implementation of the detach algorithm directly with the source file from Adam Vert's `nimbaldetach`

### Tasks

- [ ] refactor `_nimbaldetach.py`

### Freeform Notes

_No response_
