### 2025-09-11 20fd3a6
Starting refactor
