+++ src/wristpy/processing/metrics.py
@@ -7,7 +7,7 @@ import polars as pl
 from scipy import interpolate, signal
 
 from wristpy.core import computations, config, models
-from wristpy.processing import mims, nimbaldetach
+from wristpy.processing import _nimbaldetach, mims
 
 logger = config.get_logger()
 
@@ -421,7 +421,7 @@ def detach_nonwear(
         1 / ((acceleration.time[1:] - acceleration.time[:-1]).median()).total_seconds()  # type: ignore[union-attr] #Guarded by Measurement validation for .time attribute
     )
 
-    [deatch_wear, nonwear_array] = nimbaldetach.nimbaldetach(
+    [deatch_wear, nonwear_array] = _nimbaldetach.detach(
         x_values=acceleration.measurements[:, 0],
         y_values=acceleration.measurements[:, 1],
         z_values=acceleration.measurements[:, 2],
