### 2025-09-11 b752d9a
Bugfix/resolve security issues (#247)
* Update min actfast version

* Add action commit hash

* bump dependencies

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update README.md

typo? wonder how long that has been there

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
