+++ .github/workflows/docs.yaml
@@ -30,6 +30,6 @@ jobs:
       run: |
         poetry run pdoc src/wristpy -o docs_build -t docs/pdoc-theme --docformat google
         touch docs_build/.nojekyll
-    - uses: JamesIves/github-pages-deploy-action@v4
+    - uses: JamesIves/github-pages-deploy-action@6c2d9db40f9296374acc17b90404b6e8864128c8
       with:
         folder: docs_build
