+++ pyproject.toml
@@ -18,7 +18,7 @@ python = ">=3.10, <3.13"
 polars = "^1.21.0"
 scikit-learn = "^1.4.1.post1"
 pyarrow = ">=18,<20"
-actfast = "^1.1.0"
+actfast = "^1.1.3"
 numpy = ">=1.26.0, <3.0.0"
 pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
