+++ .github/workflows/test.yaml
@@ -19,7 +19,7 @@ jobs:
         python_version: ['3.10', '3.11', '3.12']
         resolution: [lowest-direct, highest]
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v5
     - name: Install poetry
       run: pipx install poetry
     - uses: actions/setup-python@v5
@@ -59,7 +59,7 @@ jobs:
   ruff:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v5
     - name: Install poetry
       run: pipx install poetry
     - uses: actions/setup-python@v5
@@ -77,7 +77,7 @@ jobs:
   mypy:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
+    - uses: actions/checkout@v5
     - name: Install poetry
       run: pipx install poetry
     - uses: actions/setup-python@v5
