### 2025-09-11 ebc5aff
Update test.yaml
Add 3.13 back
