### 2025-09-15 #8637 @sappelhoff
@LukasAdamowicz, @luciorq

Hope you are doing alright. This is another short reminder to please let us know with reference to https://github.com/openjournals/joss-reviews/issues/8637#issuecomment-3237601228 ☝️
