### 2025-09-16 #8637 @Asanto32
Hi all, I would just like to mention the following issues we are working on that should improve QoL, documentation, and cross-platform stability for 3.13, they should be completed soon and then merged into the paper branch:
https://github.com/childmindresearch/wristpy/issues/248
https://github.com/childmindresearch/wristpy/issues/218
