### 2025-09-16 275899a
refactor intermediary steps
