+++ src/wristpy/processing/_nimbaldetach.py
@@ -1,6 +1,7 @@
 """Module for DETACH non-wear algorithm.
 
-This module is taken from Adam Vert https://github.com/nimbal/nimbaldetach
+This module is taken, and then edited, from Adam Vert:
+https://github.com/nimbal/nimbaldetach
 and is licensed under the MIT License:
 
 MIT License
@@ -26,20 +27,15 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-from typing import Optional
-
 import numpy as np
-import pandas as pd
+import polars as pl
 from scipy import signal
 
 
 def detach(
-    x_values: np.ndarray,
-    y_values: np.ndarray,
-    z_values: np.ndarray,
+    acceleration_values: np.ndarray,
     temperature_values: np.ndarray,
-    accel_freq: float = 75,
-    temperature_freq: float = 0.25,
+    sampling_rate: float,
     std_thresh: float = 0.008,
     low_temperature_cutoff: float = 26,
     high_temperature_cutoff: float = 30,
@@ -53,13 +49,13 @@ def detach(
     temperature, temperature rate of change and accelerometer standard deviation of
     3 axes to detect start and stop times of non-wear periods.
 
+    The temperature data should be upsampled to match the accelerometer data
+    sampling rate before running this function.
+
     Args:
-        x_values: Accelerometer x values.
-        y_values: Accelerometer y values.
-        z_values: Accelerometer z values.
+        acceleration_values: 3 column numpy array with x, y, and z acceleration values.
         temperature_values: temperature values.
-        accel_freq: frequency of accelerometer in Hz.
-        temperature_freq: frequency of temperature sensor in Hz.
+        sampling_rate: sampling rate for the accelerometer data, in Hz.
         std_thresh: the value which the std of an axis in the window must be below
             to trigger non-wear.
         low_temperature_cutoff: Low temperature for non-wear classification.
@@ -76,134 +72,88 @@ def detach(
         A numpy array with length of the accelerometer data marked as
             either wear (0) or non-wear (1).
     """
-    x_std_fwd = pd.Series(x_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
-    y_std_fwd = pd.Series(y_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
-    z_std_fwd = pd.Series(z_values)[::-1].rolling(round(accel_freq * 60)).std()[::-1]
+    window_size_1min = int(sampling_rate * 60)
 
-    x_std_back = pd.Series(x_values).rolling(round(accel_freq * 60)).std()
-    y_std_back = pd.Series(y_values).rolling(round(accel_freq * 60)).std()
-    z_std_back = pd.Series(z_values).rolling(round(accel_freq * 60)).std()
-
-    std_df = pd.DataFrame(
-        {
-            "x_std_fwd": x_std_fwd,
-            "y_std_fwd": y_std_fwd,
-            "z_std_fwd": z_std_fwd,
-            "x_std_back": x_std_back,
-            "y_std_back": y_std_back,
-            "z_std_back": z_std_back,
-        }
+    std_df = calculate_rolling_std_polars(
+        acceleration=acceleration_values,
+        sampling_rate=sampling_rate,
+        num_axes=num_axes,
+        std_thresh=std_thresh,
+        window_size_1min=window_size_1min,
     )
 
-    smoothed_temperature = _lowpass_filter_signal(
-        temperature_values, low_f=0.005, sample_f=temperature_freq
+    smoothed_temperature = lowpass_filter_signal(
[…]
