### 2025-09-16 53e7b8d
Update _nimbaldetach.py
