### 2025-09-16 a093200
Update test.yaml
Should no longer be needed due to removing skdh dependency
