### 2025-09-16 fee7b7f
Update README.md
Remove libomp warning
