### 2025-09-16 a204efc
Update _nimbaldetach.py
change function name
