### 2025-09-17 7097904
remove old tutorial
