### 2025-09-17 3eb5612
removing old files, updating settings.
