+++ docs/source/api/wristpy.core.cli.rst
@@ -3,20 +3,21 @@
 
 .. automodule:: wristpy.core.cli
 
-
+   
    .. rubric:: Functions
 
    .. autosummary::
-
+   
       main
       version_check
-
+   
    .. rubric:: Classes
 
    .. autosummary::
-
+   
       ActivityMetric
       Calibrator
       Enum
       NonwearAlgorithms
       OutputFileType
+   
