+++ docs/source/api/wristpy.processing.calibration.rst
@@ -3,18 +3,18 @@
 
 .. automodule:: wristpy.processing.calibration
 
-
+   
    .. rubric:: Functions
 
    .. autosummary::
-
+   
       cast
       dataclass
-
+   
    .. rubric:: Classes
 
    .. autosummary::
-
+   
       AbstractCalibrator
       CalibrationDispatcher
       ConstrainedMinimizationCalibration
@@ -22,3 +22,4 @@
       GgirCalibration
       LinearTransformation
       datetime
+   
