+++ docs/source/api/wristpy.core.exceptions.rst
@@ -3,14 +3,15 @@
 
 .. automodule:: wristpy.core.exceptions
 
-
+   
    .. rubric:: Exceptions
 
    .. autosummary::
-
+   
       CalibrationError
       EmptyDirectoryError
       InvalidFileTypeError
       LoggedException
       NoMotionError
       SphereCriteriaError
+   
