+++ docs/source/index.rst
@@ -1,10 +1,17 @@
-wristpy Documentation
+Wristpy Documentation
 =====================
 
-.. image:: _static/images/wristpy_logo.png
-   :alt: wristpy logo
+.. image:: _static/images/wristpy_logo_light.png
+   :alt: wristpy logo light
    :align: center
    :width: 200px
+   :class: only-light
+
+.. image:: _static/images/wristpy_logo_dark.png
+   :alt: wristpy logo dark
+   :align: center
+   :width: 200px
+   :class: only-dark
 
 |DOI| |Build| |codecov| |Ruff| |stability| |License| |docs|
 
@@ -28,144 +35,50 @@ wristpy Documentation
 .. |docs| image:: https://img.shields.io/badge/api-docs-blue
    :target: https://childmindresearch.github.io/wristpy
 
-Welcome to wristpy, a Python library designed for processing and analyzing wrist-worn accelerometer data.
+Welcome to **wristpy**, a Python library for processing and analyzing wrist-worn accelerometer data.
 
-This library provides a set of tools for loading sensor information, calibrating raw accelerometer data, calculating various physical activity metrics, finding non-wear periods, and detecting sleep periods (onset and wakeup times). Additionally, we provide access to other sensor data that may be recorded by the watch, including; temperature, luminosity, capacitive sensing, battery voltage, and all metadata.
+Wristpy provides tools for:
 
-Quick Start
------------
+- Loading and calibrating raw accelerometer data
+- Computing physical activity levels using several established metrics (ENMO, MAD, Activity- Counts, MIMS)
+- Detecting non-wear periods and sleep onset/wake times
+- Accessing additional sensor streams (temperature, luminosity, capacitive sense, battery, metadata)
 
-.. note::
-   **⚠️ Important Note for macOS Users**
+For installation instructions and a quick introduction, see :doc:`getting_started`.
 
-   **wristpy** depends on ``libomp``, a system-level dependency that is not always installed by default on macOS.
-   Install it via:
+How To Cite 
+--------------
+If you use wristpy in your research, please cite the following paper:
 
-   .. code-block:: bash
+Santorelli A, Perez F, Vos de Wael R, Rupprecht F, d'Antonio-Bertagnolli JV,  
+Franco A, Kiar G. *wristpy: Fast, User-Friendly Python Processing of Wrist-worn Accelerometer Data*. 2025. Under review. DOI: https://doi.org/10.5281/zenodo.13883190
 
-      brew install libomp
-
-**Installation**
-
-Install wristpy from PyPI:
-
-.. code-block:: bash
-
-   pip install wristpy
-
-**Interfaces**
-
-``wristpy`` provides three flexible interfaces:
-
-- **Command-line tool** for direct execution
-- **Python library** importable in scripts or notebooks
-- **Docker image** for containerized deployment
-
-**Using the Command-Line Interface**
-
-Run a single file:
-
-.. code-block:: bash
-
-   wristpy /input/file/path.gt3x -o /save/path/file_name.csv -c gradient
-
-Run an entire directory:
-
-.. code-block:: bash
-
-   wristpy /path/to/files/input_dir -o /path/to/files/output_dir -c gradient -O .csv
-
-See all available command-line arguments:
-
-.. code-block:: bash
-
-   wristpy --help
-
-**Using Python (script or notebook)**
-
-Process a single file:
-
-.. code-block:: python
-
-   from wristpy.core import orchestrator
-
-   results = orchestrator.run(
-       input='/path/to/your/file.gt3x',
[…]
