+++ docs/source/api/wristpy.core.computations.rst
@@ -3,12 +3,13 @@
 
 .. automodule:: wristpy.core.computations
 
-
+   
    .. rubric:: Functions
 
    .. autosummary::
-
+   
       moving_mean
       moving_median
       moving_std
       resample
+   
