+++ docs/source/api/wristpy.io.readers.rst
@@ -3,7 +3,7 @@
 
 .. automodule:: wristpy.io.readers
 
-
+   
 .. rubric:: Modules
 
 .. autosummary::
