+++ docs/source/api/wristpy.core.models.rst
@@ -3,17 +3,18 @@
 
 .. automodule:: wristpy.core.models
 
-
+   
    .. rubric:: Functions
 
    .. autosummary::
-
+   
       field_validator
-
+   
    .. rubric:: Classes
 
    .. autosummary::
-
+   
       BaseModel
       Measurement
       WatchData
+   
