### 2025-09-17 541db1c
ignore notebooks
