### 2025-09-17 fb19bd0
Update _nimbaldetach.py
Fixing some docstrings
