### 2025-09-18 24c969d
Remove pandas
