+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "wristpy"
-version = "0.2.0"
+version = "0.2.1"
 description = "wristpy is a Python package designed for processing and analyzing wrist-worn accelerometer data."
 authors = [
   "Adam Santorelli <adam.santorelli@childmind.org>",
@@ -24,7 +24,6 @@ pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.14.1"
 typer = "^0.15.2"
-pandas = "^2.3.2"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.0"
