### 2025-09-18 6300eec
Update dependency, adding permissions (#244)
* Update dependency, adding permissions

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update test.yaml

* Add action commit hash

* Update min actfast version

* Update pyproject.toml

bump min version of scipy

* Update poetry.lock

* Update test.yaml

Try moving this up

* Update pyproject.toml

Move to numpy > 2.1.0?

* Update poetry.lock

* Bump scikit learn

* More scikit bump

* Bump min actfast to newest

* Update test.yaml

Fix permission level

* Initial test of nimbledetach

fai
[…]
