+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "wristpy"
-version = "0.2.0"
+version = "0.2.1"
 description = "wristpy is a Python package designed for processing and analyzing wrist-worn accelerometer data."
 authors = [
   "Adam Santorelli <adam.santorelli@childmind.org>",
@@ -14,16 +14,15 @@ readme = "README.md"
 packages = [{include = "wristpy", from = "src"}]
 
 [tool.poetry.dependencies]
-python = ">=3.10, <3.13"
+python = ">=3.10, <3.14"
 polars = "^1.21.0"
-scikit-learn = "^1.4.1.post1"
+scikit-learn = "^1.7.1"
 pyarrow = ">=18,<20"
 actfast = "^1.1.3"
-numpy = ">=1.26.0, <3.0.0"
+numpy = ">=2.1.0, <3.0.0"
 pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
-scipy = "^1.13.0"
-scikit-digital-health = "^0.17.9"
+scipy = "^1.14.1"
 typer = "^0.15.2"
 
 [tool.poetry.group.dev.dependencies]
