+++ .github/workflows/test.yaml
@@ -9,28 +9,26 @@ on:
     - main
   pull_request:
 
+
 jobs:
   unit:
     runs-on: ${{ matrix.os }}
+
     strategy:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        python_version: ['3.10', '3.11', '3.12']
+        python_version: ['3.10', '3.11', '3.12', '3.13']
         resolution: [lowest-direct, highest]
     steps:
     - uses: actions/checkout@v4
     - name: Install poetry
       run: pipx install poetry
+
     - uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python_version }}
         cache: pip
-
-    - name: Install dependencies on macOS
-      if: runner.os == 'macOS'
-      run: |
-        brew install libomp gfortran
 
     - name: Install dependencies
       run: |
