+++ pyproject.toml
@@ -29,8 +29,8 @@ typer = "^0.15.2"
 pytest = "^8.0.0"
 mypy = "^1.14.1"
 pre-commit = "^4.0.1"
-pytest-cov = ">=5,<7"
-ruff = ">=0.7.1,<0.9.0"
+pytest-cov = ">=5,<8"
+ruff = ">=0.7.1,<0.14.0"
 pytest-mock = "^3.14.0"
 
 [tool.poetry.group.docs.dependencies]
