+++ docs/source/api/wristpy.processing.processing_utils.rst
@@ -3,11 +3,11 @@
 
 .. automodule:: wristpy.processing.processing_utils
 
-   
+
    .. rubric:: Functions
 
    .. autosummary::
-   
+
       get_nonwear_measurements
       synchronize_measurements
-   
+
