+++ docs/source/api/wristpy.processing.metrics.rst
@@ -3,11 +3,11 @@
 
 .. automodule:: wristpy.processing.metrics
 
-   
+
    .. rubric:: Functions
 
    .. autosummary::
-   
+
       actigraph_activity_counts
       angle_relative_to_horizontal
       combined_temp_accel_detect_nonwear
@@ -16,4 +16,4 @@
       euclidean_norm_minus_one
       mean_amplitude_deviation
       monitor_independent_movement_summary_units
-   
+
