### 2025-09-18 660a872
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
