+++ docs/source/api/wristpy.processing.mims.rst
@@ -13,4 +13,3 @@
       combine_mims
       extrapolate_points
       interpolate_measure
-
