+++ docs/source/api/wristpy.processing.idle_sleep_mode_imputation.rst
@@ -9,4 +9,3 @@
    .. autosummary::
 
       impute_idle_sleep_mode_gaps
-
