### 2025-09-18 06988e7
removing type ignores, updating toml to ignore notebook.
