### 2025-09-19 dbf9df1
Update .github/workflows/docs.yaml
Co-authored-by: Adam Santorelli <148909356+Asanto32@users.noreply.github.com>
