+++ .github/workflows/docs.yaml
@@ -4,11 +4,12 @@ permissions:
   contents: write
 
 on:
-  push:
+  workflow_run:
+    workflows: [Python Tests]
+    types:
+    - completed
     branches:
-    - feature/issue-218/sphinx
     - main
-  workflow_dispatch:
 
 jobs:
   build-and-publish-docs:
@@ -31,7 +32,7 @@ jobs:
         touch build/html/.nojekyll
         ls -la build/html | sed 's/^/BUILD: /'
     - name: Deploy to gh-pages
-      uses: JamesIves/github-pages-deploy-action@c2d9db40f9296374acc17b90404b6e8864128c8
+      uses: JamesIves/github-pages-deploy-action@6c2d9db40f9296374acc17b90404b6e8864128c8
       with:
         branch: gh-pages
         folder: docs/build/html
