+++ pyproject.toml
@@ -32,7 +32,7 @@ pre-commit = "^4.0.1"
 pytest-cov = ">=5,<7"
 ruff = ">=0.7.1,<0.9.0"
 pytest-mock = "^3.14.0"
-sphinx-autobuild = "^2024.2.4"
+
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.0.0"
@@ -42,6 +42,7 @@ myst-nb = "^1.3"
 sphinx-copybutton = "^0.5.2"
 sphinx-design = "^0.5.0"
 ipykernel = "^6.29.0"
+sphinx-autobuild = "^2024.2.4"
 
 [tool.poetry.scripts]
 wristpy = "wristpy.__main__:run_main"
