+++ docs/Makefile
@@ -17,7 +17,11 @@ help:
 
 
 clean:
-	rm -rf $(BUILDDIR)/*
+	 @if [ -z "$(BUILDDIR)" ]; then \
+        echo "Error: BUILDDIR is not set."; \
+        exit 1; \
+    fi
+    rm -rf $(BUILDDIR)/*
 
 livehtml:
 	sphinx-autobuild "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
