### 2025-09-23 1ff7d9a
Feature/issue 218/sphinx (#238)
* new docs

* api change

* more api automation changes.

* docs workflow

* build from branch

* link

* rm pdocs

* command change and duplicate code.

* deploy from gh-pages

* push to deploy

* actions change

* minimal actions, build on push

* single commit

* __doc__ link

* tutorial change

* homepage change.

* git ignore.

* reverting workflows to work off main.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* mistake.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more informatio
[…]
