+++ docs/source/api/wristpy.io.readers.rst
@@ -1,0 +1,13 @@
+﻿wristpy.io.readers
+==================
+
+.. automodule:: wristpy.io.readers
+
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree:
+   :recursive:
+
+   readers
