+++ docs/wristpy_tutorial.md
@@ -1,275 +1,0 @@
-# Wristpy Tutorial
-
-
-## Introduction
-
-
-Wristpy is a Python library designed for processing and analyzing wrist-worn accelerometer data.
-This tutorial will guide you through the basic steps of using `wristpy` to analyze your accelerometer data. Specifically,
-we will cover the following topics through a few examples:
-   - running the default processor, analyzing the output data, and visualizing the results.
-   - loading data and plotting the raw signals.
-   - how to calibrate the data, computing ENMO and angle-z from the calibrated data and then plotting those metrics.
-   - how to obtain non-wear windows and visualize them.
-   - how to obtain sleep windows and visualize them.
-
-
-
-
-
-## Example 1: Running the default processor
-
-
-The `orchestrator` module of wristpy contains the default processor that will run the entire wristpy processing pipeline. This can be called as simply as:
-
-```python
-from wristpy.core import orchestrator
-
-results = orchestrator.run(
-   input = '/path/to/your/file.gt3x',
-   output = 'path/to/save/file_name.csv'
-)
-```
-This runs the processing pipeline with all the default arguments, creates an output `.csv` file, a `.json` file with the pipeline configuration parameters, and will create a `results` object that contains the various output metrics (namely; the specified physical activity metric, angle-z, physical activity classification values, non-wear status, and sleep status).
-
-
-The orchestrator can also process entire directories. The call to the orchestrator remains largely the same but now output is expected to be a directory and the desired filetype for the saved files can be specified through the output_filetype arguement(default value is ".csv"):
-
-```python
-from wristpy.core import orchestrator
-
-results = orchestrator.run(
-    input = '/path/to/input/dir',
-    output = '/path/to/output/dir',
-    output_filetype = ".csv"
-)
-```
-
-If users would prefer to process specific files instead of entire directories we recommend looping through a list of file names. The following code snipet will save results objects into a dictionary, and the output files into the desired directory:
-
-```python
-from wristpy.core import orchestrator
-import pathlib
-
-file_path = pathlib.Path("/path/to/data/")
-output_dir = pathlib.Path("/path/to/save/dir/")
-file_names = [pathlib.Path("file1.gt3x"), pathlib.Path("file2.gt3x"), pathlib.Path("file3.gt3x")]
-results_dict = {}
-
-for file in file_names:
-    input_path = file_path / file
-    output_path = output_dir / file.stem + ".csv"
-    result = orchestrator.run(
-        input = input_path
-        output = output_path
-    )
-    results_dict[file.stem] = result
-```
-
-
-
-
-We can visualize some of the outputs within the `results` object, directly, with the following scripts:
-
-Plot the default physical activity metrics (ENMO) across the entire data set:
-```python
-from matplotlib import pyplot as plt
-plt.plot(results.physical_activity_metric.time, results.physical_activity_metric.measurements)
-```
-
-![Example of the ENMO result](enmo_example1.png)
-
-Plot the sleep windows with normalized angle-z data:
-```python
-from matplotlib import pyplot as plt
-
-plt.plot(results.anglez.time, results.anglez.measurements/90)
-plt.plot(results.sleep_status.time, results.sleep_status.measurements)
-plt.legend(['Angle Z', 'Sleep Windows'])
-plt.show()
-```
-![Example of the Sleep and Anglez](sleep_anglez_example1.png)
-
[…]
