+++ docs/source/api/wristpy.core.computations.rst
@@ -1,0 +1,14 @@
+﻿wristpy.core.computations
+=========================
+
+.. automodule:: wristpy.core.computations
+
+
+   .. rubric:: Functions
+
+   .. autosummary::
+
+      moving_mean
+      moving_median
+      moving_std
+      resample
