+++ pyproject.toml
@@ -34,7 +34,14 @@ ruff = ">=0.7.1,<0.9.0"
 pytest-mock = "^3.14.0"
 
 [tool.poetry.group.docs.dependencies]
-pdoc = "^15.0.0"
+sphinx = "^7.0.0"
+furo = "^2024.1.0"
+myst-parser = "^2.0.0"
+myst-nb = "^1.3"
+sphinx-copybutton = "^0.5.2"
+sphinx-design = "^0.5.0"
+ipykernel = "^6.29.0"
+sphinx-autobuild = "^2024.2.4"
 
 [tool.poetry.scripts]
 wristpy = "wristpy.__main__:run_main"
@@ -49,6 +56,7 @@ testpaths = [
 
 [tool.mypy]
 ignore_missing_imports = true
+exclude = ["docs/"]
 
 [tool.ruff]
 exclude = [
@@ -72,7 +80,8 @@ exclude = [
   "build",
   "dist",
   "node_modules",
-  "venv"
+  "venv",
+  "*.ipynb"
 ]
 line-length = 88
 indent-width = 4
@@ -100,6 +109,7 @@ convention = "google"
 
 [tool.ruff.per-file-ignores]
 "tests/**/*.py" = []
+"docs/source/conf.py" = ["INP001"]
 
 [build-system]
 requires = ["poetry-core>=1.2.0"]
