+++ docs/Makefile
@@ -1,0 +1,27 @@
+
+
+
+SPHINXOPTS    ?=
+SPHINXBUILD  ?= sphinx-build
+SOURCEDIR    = source
+BUILDDIR     = build
+
+
+help:
+	@$(SPHINXBUILD) -M help "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
+
+.PHONY: help Makefile
+
+%: Makefile
+	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
+
+
+clean:
+	 @if [ -z "$(BUILDDIR)" ]; then \
+        echo "Error: BUILDDIR is not set."; \
+        exit 1; \
+    fi
+    rm -rf $(BUILDDIR)/*
+
+livehtml:
+	sphinx-autobuild "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)

[…]
