### 2025-09-26 childmindresearch/wristpy#248 @Asanto32
This was addressed in #244
