### 2025-10-01 f9873dc
Bump the production-dependencies group across 1 directory with 3 updates
Bumps the production-dependencies group with 3 updates in the / directory: [pyarrow](https://github.com/apache/arrow), [pydantic-settings](https://github.com/pydantic/pydantic-settings) and [typer](https://github.com/fastapi/typer).


Updates `pyarrow` from 19.0.1 to 21.0.0
- [Release notes](https://github.com/apache/arrow/releases)
- [Commits](https://github.com/apache/arrow/compare/apache-arrow-19.0.1...apache-arrow-21.0.0)

Updates `pydantic-settings` from 2.10.1 to 2.11.0
- [Release notes](https://github.com/pydantic/
[…]
