### 2025-10-01 e4ac8f3
Bump the development-dependencies group across 1 directory with 6 updates
Bumps the development-dependencies group with 6 updates in the / directory:

| Package | From | To |
| --- | --- | --- |
| [mypy](https://github.com/python/mypy) | `1.18.1` | `1.18.2` |
| [pytest-cov](https://github.com/pytest-dev/pytest-cov) | `6.3.0` | `7.0.0` |
| [ruff](https://github.com/astral-sh/ruff) | `0.8.6` | `0.13.1` |
| [furo](https://github.com/pradyunsg/furo) | `2024.8.6` | `2025.7.19` |
| [myst-parser](https://github.com/executablebooks/MyST-Parser) | `2.0.0` | `4.0.1` |
| [sphinx-design](https://github.co
[…]
