+++ pyproject.toml
@@ -29,17 +29,17 @@ typer = "^0.15.2"
 pytest = "^8.0.0"
 mypy = "^1.14.1"
 pre-commit = "^4.0.1"
-pytest-cov = ">=5,<7"
-ruff = ">=0.7.1,<0.9.0"
+pytest-cov = ">=5,<8"
+ruff = ">=0.7.1,<0.14.0"
 pytest-mock = "^3.14.0"
 
 [tool.poetry.group.docs.dependencies]
 sphinx = "^7.0.0"
-furo = "^2024.1.0"
-myst-parser = "^2.0.0"
+furo = ">=2024.1,<2026.0"
+myst-parser = ">=2,<5"
 myst-nb = "^1.3"
 sphinx-copybutton = "^0.5.2"
-sphinx-design = "^0.5.0"
+sphinx-design = ">=0.5,<0.7"
 ipykernel = "^6.29.0"
 sphinx-autobuild = "^2024.2.4"
 
