### 2025-10-02 9f465ea
Add time zone metadata
Different metadata structure for gt3x and bin
