### 2025-10-03 d16447d
PR review modifications
Added IOError catch into readers
Remove GMT from timezone string
