### 2025-10-03 7351f8e
Change to ValueError for bad extension
