### 2025-10-03 8b25392
Add time zone metadata (#259)
* Add time zone metadata

Different metadata structure for gt3x and bin

* Bump actions checkout

* PR review modifications

Added IOError catch into readers
Remove GMT from timezone string

* Change to ValueError for bad extension
