### 2025-10-07 6a2dc80
Add progress bar
Add progress bar, poetry update, Typo in INFO message for directory that is processing
