+++ pyproject.toml
@@ -24,6 +24,7 @@ pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.14.1"
 typer = "^0.15.2"
+rich = "^14.1.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.0.0"
