### 2025-10-07 631ca69
Add progress bar (#260)
* Add progress bar

Add progress bar, poetry update, Typo in INFO message for directory that is processing

* Edits for PR
