### 2025-10-07 ccbbc57
Bump the production-dependencies group across 1 directory with 3 updates
Bumps the production-dependencies group with 3 updates in the / directory: [pyarrow](https://github.com/apache/arrow), [pydantic](https://github.com/pydantic/pydantic) and [typer](https://github.com/fastapi/typer).


Updates `pyarrow` from 19.0.1 to 21.0.0
- [Release notes](https://github.com/apache/arrow/releases)
- [Commits](https://github.com/apache/arrow/compare/apache-arrow-19.0.1...apache-arrow-21.0.0)

Updates `pydantic` from 2.11.10 to 2.12.0
- [Release notes](https://github.com/pydantic/pydantic/releases)
- [Chan
[…]
