+++ pyproject.toml
@@ -17,13 +17,13 @@ packages = [{include = "wristpy", from = "src"}]
 python = ">=3.10, <3.14"
 polars = "^1.21.0"
 scikit-learn = "^1.7.1"
-pyarrow = ">=18,<20"
+pyarrow = ">=18,<22"
 actfast = "^1.1.3"
 numpy = ">=2.1.0, <3.0.0"
 pydantic = "^2.11.1"
 pydantic-settings = "^2.3.4"
 scipy = "^1.14.1"
-typer = "^0.15.2"
+typer = ">=0.15.2,<0.20.0"
 rich = "^14.1.0"
 
 [tool.poetry.group.dev.dependencies]
