### 2025-10-16 bbbe70c
Update paper.md
