+++ paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 
 # Summary
 
-`wristpy` is an open-source Python package designed to streamline the processing and analysis of wrist-worn accelerometer data. The package has been developed in a modular framework to process actigraphy data from both GENEActiv and Actigraph watches, while supporting the extension for other wearable standards in the future, and is supported across multiple platforms (Windows, macOS, and Ubuntu). This software can be accessed and used through either a command-line interface (CLI) or as an importable Python module. The main processing pipeline will generate outputs in either a `.csv` or `.parquet` format. The output contains timeseries data at a user-chosen temporal resolution, including: physical activity metrics, sleep onset and wake up times, non-wear period identification, and physical activity classification.
+`wristpy` is an open-source Python package designed to streamline the processing and analysis of wrist-worn accelerometer data. The package has been developed in a modular framework to process actigraphy data from both GENEActiv and Actigraph watches, while supporting the extension for other wearable standards in the future, and is supported across multiple platforms (Windows, macOS, and Ubuntu). This software can be accessed and used through either a command-line interface (CLI) or as an importable Python module. The main processing pipeline will generate outputs in either a `.csv` or `.parquet` format. The output contains timeseries data at a user-chosen temporal resolution, including: physical activity metrics, sleep status (including the sustained inactivity bouts (SIB) and the sleep period time (SPT) windows), non-wear period identification, and physical activity classification.
 
 
 `wristpy` has been designed modularly, which allows for incremental improvements of bottlenecks in the processing pipeline — such as the independent creation and adoption of a Rust-based reader for actigraphy data that dramatically improves scalability over the vendor-provided reader. With `wristpy`, researchers can directly access a variety of functions that allow them to read specific sensor data stored on the watch, calculate a variety of physical activity metrics, compare various non-wear or sleep detection algorithms, or take advantage of them together in a 1-click end-to-end preprocessing pipeline.
@@ -46,9 +46,9 @@ bibliography: paper.bib
 
 Wearable accelerometers are increasingly used to measure both physical activity and sleep patterns as they provide a simple, effective, non-invasive, and low-cost alternative to clinical observation, allowing for large scale data to be acquired in realistic scenarios. While several commercial solutions exist, such as consumer-grade products providing aggregate data (e.g., FitBit) or research-focused proprietary software (e.g., such as those from Actigraph and ActivInsights), a need exists for device-agnostic open-source tools that can process raw accelerometer data with transparency, reproducibility, and scalability, both as a base for evaluation and methodological advancement.
 
-Several open-source packages have been developed, namely GGIR [@van_hees_ggir_2025], Scikit Digital Health (`skdh`) [@adamowicz_scikit_2022], and pyActigraphy [@hammad_pyactigraphy_2021]. GGIR is one of the most commonly used software packages to process and analyze actigraphy data from various watches and has been cited in hundreds of publications over the past decade. `skdh` is a Python package that can process data from GENEActiv watches to compute physical activity metrics, sleep detection based on [@christakis_sleeppy_2019], novel wear detection [@vert_detecting_2022], gait detection, and numerous other functions for both processing and analyzing accelerometer data. 
[…]
