### 2025-10-20 #8637 @Asanto32
Hi @sappelhoff,

I have finished the edits to `paper.md`, and completed the below checklist (see inline comments for details).
Thanks and please let me know if you need anything else from me!

> ## Additional Author Tasks After Review is Complete

>  [x] Double check authors and affiliations (including ORCIDs).

Completed.

>   Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

The latest version on main is v0.2.2 and contains numerous improvements that have been outlined in our discussions here.

> [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.

The latest (to ensure author list compliance) Zenodo archive is:  [10.5281/zenodo.17289865](https://doi.org/10.5281/zenodo.17289865)

> [x]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

> [x]  Make sure that the license listed for the archive is the same as the software license.
