### 2025-10-21 879fbf1
Update paper.md
