### 2025-10-23 218d843
Feature/258 tutorial breakup (#263)
* Getting started breakup

* menu, fixing cell outputs.

* spell checking

* Update README.md

typo

* Typo/description fix

---------

Co-authored-by: Adam Santorelli <148909356+Asanto32@users.noreply.github.com>
