### 2025-10-28 d6ded55
Update paper.md
Fix typos + affiliations
