### 2025-10-28 #8637 @Asanto32
Hi @Kevin-Mattheus-Moerman,

Apologies for those typos. Both typos and the affiliation spelling of USA should now be complete.
